working on docs! [refactor][gui] split out model inputs from model widget [refactor][docs] - split model widget file, moved examples to folder [docs] - cleaning up the configuration, training, and task guide
