--- docs/data/index.rst
@@ -1,16 +1,0 @@
-data
-====
-
-.. automodule:: autopilot.data
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-.. toctree::
-
-    subject
-    interfaces/index
-    modeling/index
-    models/index
-    units/index
