--- docs/examples.rst
@@ -1,13 +1,0 @@
-Examples
-=========
-
-We're working on writing more examples! Please let us know in the discussion board what you'd like to see :)
-
-Also see the ``examples`` folder in the repository for jupyter notebooks we haven't set up Sphinx rendering for yet ;)
-
-
-.. toctree::
-    maxdepth: 1
-    :caption: Tasks
-
-    examples.tasks.blink
