+++ docs/examples/blink.rst
@@ -169,7 +169,7 @@ message would only appear if ``LOGLEVEL == "DEBUG"``::
 We set the stage block and never clear it so that the :class:`.Pilot` doesn't wait for a trigger
 to call the next stage -- it just does it as soon as the previous one completes.
 
-See :meth:`~autopilot.core.pilot.Pilot.run_task` for more detail on this loop.
+See :meth:`~autopilot.agents.pilot.Pilot.run_task` for more detail on this loop.
 
 Stage Methods
 --------------
@@ -225,6 +225,6 @@ a column for each of the fields here.
 Full Source
 -------------
 
-.. literalinclude:: ../examples/tasks/blink.py
+.. literalinclude:: ../../examples/tasks/blink.py
     :language: python
     :linenos:
