+++ docs/gui/widgets.rst
@@ -1,49 +1,9 @@
-widgets
-=========
-
-.. automodule:: autopilot.gui.widgets
-    :members:
-    :undoc-members:
-    :show-inheritance:
 
 
-model
-------
-
-.. automodule:: autopilot.gui.widgets.model
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-protocol
----------
-
-.. automodule:: autopilot.gui.widgets.model
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-subject
---------
-
-.. automodule:: autopilot.gui.widgets.subject
-    :members:
-    :undoc-members:
-    :show-inheritance:
 
 
-terminal
----------
 
-.. automodule:: autopilot.gui.widgets.terminal
-    :members:
-    :undoc-members:
-    :show-inheritance:
 
-list
------
 
-.. automodule:: autopilot.gui.widgets.list
-    :members:
-    :undoc-members:
-    :show-inheritance:
+
+
