+++ docs/gui/widgets/index.rst
@@ -1,0 +1,16 @@
+widgets
+=========
+
+.. automodule:: autopilot.gui.widgets
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+.. toctree::
+
+    input
+    model
+    protocol
+    subject
+    terminal
+    list
