+++ docs/gui/widgets/input.rst
@@ -1,0 +1,6 @@
+input
+------
+
+.. automodule:: autopilot.gui.widgets.input
+    :members:
+    :autosummary-no-nesting:
