+++ docs/gui/widgets/list.rst
@@ -1,0 +1,7 @@
+list
+-----
+
+.. automodule:: autopilot.gui.widgets.list
+    :members:
+    :undoc-members:
+    :show-inheritance:
