+++ docs/gui/widgets/model.rst
@@ -1,0 +1,5 @@
+model
+------
+
+.. automodule:: autopilot.gui.widgets.model
+    :members:
