+++ docs/gui/widgets/protocol.rst
@@ -1,0 +1,8 @@
+
+protocol
+---------
+
+.. automodule:: autopilot.gui.widgets.model
+    :members:
+    :undoc-members:
+    :show-inheritance:
