+++ docs/gui/widgets/subject.rst
@@ -1,0 +1,7 @@
+subject
+--------
+
+.. automodule:: autopilot.gui.widgets.subject
+    :members:
+    :undoc-members:
+    :show-inheritance:
