+++ docs/gui/widgets/terminal.rst
@@ -1,0 +1,7 @@
+terminal
+---------
+
+.. automodule:: autopilot.gui.widgets.terminal
+    :members:
+    :undoc-members:
+    :show-inheritance:
