+++ docs/searchindex.js
@@ -1,1 +1,1 @@
-Search.setIndex({docnames:["autopilot.core","autopilot.agents.gui","autopilot.core.networking","autopilot.core.pilot","autopilot.agents.plots","autopilot.agents.styles","autopilot.core.subject","autopilot.core.terminal","autopilot.core.utils","autopilot.external","autopilot.hardware","autopilot.hardware.cameras","autopilot.hardware.gpio","autopilot.hardware.i2c","autopilot.hardware.usb","autopilot.prefs","autopilot.setup","autopilot.setup.setup_pilot","autopilot.setup.setup_scale","autopilot.setup.setup_terminal","autopilot.stim","autopilot.stim.managers","autopilot.stim.sound","autopilot.stim.sound.jackclient","autopilot.stim.sound.pyoserver","autopilot.stim.sound.sounds","autopilot.tasks","autopilot.tasks.children","autopilot.tasks.free_water","autopilot.tasks.graduation","autopilot.tasks.nafc","autopilot.tasks.task","autopilot.viz","autopilot.viz.psychometric","autopilot.viz.trial_viewer","changelog/index","changelog/v0.2.0","changelog/v0.3.0","guide.hardware","guide.installation","guide.installation.pilot","guide.installation.terminal","guide.overview","guide.task","guide.training","index","todo"],envversion:{"sphinx.domains.c":2,"sphinx.domains.changeset":1,"sphinx.domains.citation":1,"sphinx.domains.cpp":3,"sphinx.domains.index":1,"sphinx.domains.javascript":2,"sphinx.domains.math":2,"sphinx.domains.python":2,"sphinx.domains.rst":2,"sphinx.domains.std":1,"sphinx.ext.intersphinx":1,"sphinx.ext.todo":2,"sphinx.ext.viewcode":1,sphinx:56},filenames:["autopilot.agents.rst","autopilot.agents.gui.rst","autopilot.agents.networking.rst","autopilot.agents.pilot.rst","autopilot.agents.plots.rst","autopilot.agents.styles.rst","autopilot.agents.subject.rst","autopilot.agents.terminal.rst","autopilot.agents.utils.rst","autopilot.external.rst","autopilot.hardware.rst","autopilot.hardware.cameras.rst","autopilot.hardware.gpio.rst","autopilot.hardware.i2c.rst","autopilot.hardware.usb.rst","autopilot.prefs.rst","autopilot.setup.rst","autopilot.setup.setup_pilot.rst","autopilot.setup.setup_scale.rst","autopilot.setup.setup_terminal.rst","autopilot.stim.rst","autopilot.stim.managers.rst","autopilot.stim.sound.rst","autopilot.stim.sound.jackclient.rst","autopilot.stim.sound.pyoserver.rst","autopilot.stim.sound.sounds.rst","autopilot.tasks.rst","autopilot.tasks.children.rst","autopilot.tasks.free_water.rst","autopilot.tasks.graduation.rst","autopilot.tasks.nafc.rst","autopilot.tasks.task.rst","autopilot.viz.rst","autopilot.viz.psychometric.rst","autopilot.viz.trial_viewer.rst","changelog/index.rst","changelog/v0.2.0.rst","changelog/v0.3.0.rst","guide.hardware.rst","guide.installation.rst","guide.installation.pilot.rst","guide.installation.terminal.rst","guide.overview.rst","guide.task.rst","guide.training.rst","index.rst","todo.rst"],objects:{"autopilot.core":{networking:[2,0,0,"-"],pilot:[3,0,0,"-"],styles:[5,0,0,"-"],subject:[6,0,0,"-"],utils:[8,0,0,"-"]},"autopilot.core.networking":{Message:[2,1,1,""],Net_Node:[2,1,1,""],Pilot_Station:[2,1,1,""],Station:[2,1,1,""],Terminal_Station:[2,1,1,""],serialize_array:[2,4,1,""]},"autopilot.core.networking.Message":{__contains__:[2,2,1,""],__delitem__:[2,2,1,""],__getitem__:[2,2,1,""],__setitem__:[2,2,1,""],_serialize_numpy:[2,2,1,""],changed:[2,3,1,""],expand:[2,2,1,""],flags:[2,3,1,""],get_timestamp:[2,2,1,""],id:[2,3,1,""],key:[2,3,1,""],sender:[2,3,1,""],serialize:[2,2,1,""],serialized:[2,3,1,""],timestamp:[2,3,1,""],to:[2,3,1,""],ttl:[2,3,1,""],validate:[2,2,1,""],value:[2,3,1,""]},"autopilot.core.networking.Net_Node":{context:[2,3,1,""],get_stream:[2,2,1,""],handle_listen:[2,2,1,""],id:[2,3,1,""],init_logging:[2,2,1,""],init_networking:[2,2,1,""],l_confirm:[2,2,1,""],l_stream:[2,2,1,""],listens:[2,3,1,""],log_formatter:[2,3,1,""],log_handler:[2,3,1,""],logger:[2,3,1,""],loop:[2,3,1,""],loop_thread:[2,3,1,""],outbox:[2,3,1,""],port:[2,3,1,""],prepare_message:[2,2,1,""],release:[2,2,1,""],repeat:[2,2,1,""],repeat_interval:[2,3,1,""],send:[2,2,1,""],sock:[2,3,1,""],threaded_loop:[2,2,1,""],timers:[2,3,1,""],upstream:[2,3,1,""]},"autopilot.core.networking.Pilot_Station":{l_change:[2,2,1,""],l_child:[2,2,1,""],l_cohere:[2,2,1,""],l_continuous:[2,2,1,""],l_file:[2,2,1,""],l_forward:[2,2,1,""],l_noop:[2,2,1,""],l_ping:[2,2,1,""],l_start:[2,2,1,""],l_state:[2,2,1,""],l_stop:[2,2,1,""]},"autopilot.core.networking.Station":{child:[2,3,1,""],ctx:[2,3,1,""],get_ip:[2,2,1,""],handle_listen:[2,2,1,""],id:[2,3,1,""],init_logging:[2,2,1,""],ip:[2,3,1,""],l_confirm:[2,2,1,""],l_stream:[2,2,1,""],listen_port:[2,3,1,""],listener:[2,3,1,""],listens:[2,3,1,""],log_formatter:[2,3,1,""],log_handler:[2,3,1,""],logger:[2,3,1,""],loop:[2,3,1,""],prepare_message:[2,2,1,""],push:[2,2,1,""],push_id:[2,3,1,""],push_ip:[2,3,1,""],push_outbox:[2,3,1,""],push_port:[2,3,1,""],pusher:[2,3,1,""],release:[2,2,1,""],repeat:[2,2,1,""],repeat_interval:[2,3,1,""],routes:[2,3,1,""],run:[2,2,1,""],send:[2,2,1,""],send_outbox:[2,3,1,""],senders:[2,3,1,""],timers:[2,3,1,""]},"autopilot.core.networking.Terminal_Station":{l_change:[2,2,1,""],l_continuous:[2,2,1,""],l_data:[2,2,1,""],l_file:[2,2,1,""],l_handshake:[2,2,1,""],l_init:[2,2,1,""],l_kill:[2,2,1,""],l_ping:[2,2,1,""],l_state:[2,2,1,""],l_stopall:[2,2,1,""],plot_timer:[2,3,1,""],sent_plot:[2,3,1,""],start_plot_timer:[2,2,1,""]},"autopilot.core.pilot":{Pilot:[3,1,1,""]},"autopilot.core.pilot.Pilot":{blank_LEDs:[3,2,1,""],calibrate_port:[3,2,1,""],calibration_curve:[3,2,1,""],file_block:[3,3,1,""],get_ip:[3,2,1,""],handshake:[3,2,1,""],init_audio:[3,2,1,""],init_logging:[3,2,1,""],init_pigpio:[3,2,1,""],l_bandwidth:[3,2,1,""],l_cal_port:[3,2,1,""],l_cal_result:[3,2,1,""],l_param:[3,2,1,""],l_start:[3,2,1,""],l_stop:[3,2,1,""],log_formatter:[3,3,1,""],log_handler:[3,3,1,""],logger:[3,3,1,""],networking:[3,3,1,""],node:[3,3,1,""],open_file:[3,2,1,""],run_task:[3,2,1,""],running:[3,3,1,""],server:[3,3,1,""],stage_block:[3,3,1,""],update_state:[3,2,1,""]},"autopilot.core.subject":{Subject:[6,1,1,""]},"autopilot.core.subject.Subject":{Hash_Table:[6,1,1,""],History_Table:[6,1,1,""],Weight_Table:[6,1,1,""],apply_along:[6,2,1,""],assign_protocol:[6,2,1,""],close_hdf:[6,2,1,""],data_thread:[6,2,1,""],ensure_structure:[6,2,1,""],flush_current:[6,2,1,""],get_step_history:[6,2,1,""],get_timestamp:[6,2,1,""],get_trial_data:[6,2,1,""],get_weight:[6,2,1,""],graduate:[6,2,1,""],new_subject_file:[6,2,1,""],open_hdf:[6,2,1,""],prepare_run:[6,2,1,""],save_data:[6,2,1,""],set_weight:[6,2,1,""],stash_current:[6,2,1,""],stop_run:[6,2,1,""],to_csv:[6,2,1,""],update_biography:[6,2,1,""],update_history:[6,2,1,""],update_weights:[6,2,1,""]},"autopilot.core.subject.Subject.Hash_Table":{columns:[6,3,1,""]},"autopilot.core.subject.Subject.History_Table":{columns:[6,3,1,""]},"autopilot.core.subject.Subject.Weight_Table":{columns:[6,3,1,""]},"autopilot.core.utils":{Param:[8,1,1,""],ReturnThread:[8,1,1,""],coerce_discrete:[8,4,1,""],list_subjects:[8,4,1,""],load_pilotdb:[8,4,1,""]},"autopilot.core.utils.Param":{__contains__:[8,2,1,""],__delitem__:[8,2,1,""],__getitem__:[8,2,1,""],__setitem__:[8,2,1,""],tag:[8,3,1,""],type:[8,3,1,""],types:[8,3,1,""]},"autopilot.core.utils.ReturnThread":{join:[8,2,1,""],run:[8,2,1,""]},"autopilot.hardware":{BCM_TO_BOARD:[10,5,1,""],BOARD_TO_BCM:[10,5,1,""],Hardware:[10,1,1,""],cameras:[11,0,0,"-"],gpio:[12,0,0,"-"],i2c:[13,0,0,"-"],usb:[14,0,0,"-"]},"autopilot.hardware.Hardware":{assign_cb:[10,2,1,""],get_name:[10,2,1,""],init_logging:[10,2,1,""],init_networking:[10,2,1,""],input:[10,3,1,""],is_trigger:[10,3,1,""],output:[10,3,1,""],pin:[10,3,1,""],release:[10,2,1,""],type:[10,3,1,""]},"autopilot.hardware.cameras":{Camera:[11,1,1,""],Camera_CV:[11,1,1,""],Camera_Spinnaker:[11,1,1,""],OPENCV_LAST_INIT_TIME:[11,5,1,""],Video_Writer:[11,1,1,""],list_spinnaker_cameras:[11,4,1,""]},"autopilot.hardware.cameras.Camera":{_capture:[11,2,1,""],_grab:[11,2,1,""],_process:[11,2,1,""],_timestamp:[11,2,1,""],_write_deinit:[11,2,1,""],_write_frame:[11,2,1,""],cam:[11,2,1,""],capture:[11,2,1,""],capture_deinit:[11,2,1,""],capture_init:[11,2,1,""],init_cam:[11,2,1,""],input:[11,3,1,""],l_start:[11,2,1,""],l_stop:[11,2,1,""],output_filename:[11,2,1,""],queue:[11,2,1,""],release:[11,2,1,""],stop:[11,2,1,""],stream:[11,2,1,""],type:[11,3,1,""],write:[11,2,1,""]},"autopilot.hardware.cameras.Camera_CV":{_grab:[11,2,1,""],_timestamp:[11,2,1,""],backend:[11,2,1,""],fps:[11,2,1,""],init_cam:[11,2,1,""],release:[11,2,1,""],shape:[11,2,1,""],v4l_info:[11,2,1,""]},"autopilot.hardware.cameras.Camera_Spinnaker":{ATTR_TYPES:[11,3,1,""],ATTR_TYPE_NAMES:[11,3,1,""],RW_MODES:[11,3,1,""],_grab:[11,2,1,""],_process:[11,2,1,""],_timestamp:[11,2,1,""],_write_deinit:[11,2,1,""],_write_frame:[11,2,1,""],acquisition_mode:[11,2,1,""],bin:[11,2,1,""],capture_deinit:[11,2,1,""],capture_init:[11,2,1,""],device_info:[11,2,1,""],exposure:[11,2,1,""],fps:[11,2,1,""],frame_trigger:[11,2,1,""],get:[11,2,1,""],init_cam:[11,2,1,""],list_options:[11,2,1,""],readable_attributes:[11,2,1,""],release:[11,2,1,""],set:[11,2,1,""],writable_attributes:[11,2,1,""],write:[11,2,1,""]},"autopilot.hardware.cameras.Video_Writer":{run:[11,2,1,""]},"autopilot.hardware.gpio":{Digital_In:[12,1,1,""],Digital_Out:[12,1,1,""],ENABLED:[12,5,1,""],GPIO:[12,1,1,""],INVERSE_PULL_MAP:[12,5,1,""],INVERSE_TRIGGER_MAP:[12,5,1,""],LED_RGB:[12,1,1,""],PULL_MAP:[12,5,1,""],PWM:[12,1,1,""],Solenoid:[12,1,1,""],TRIGGER_MAP:[12,5,1,""]},"autopilot.hardware.gpio.Digital_In":{assign_cb:[12,2,1,""],clear_cb:[12,2,1,""],input:[12,3,1,""],is_trigger:[12,3,1,""],record_event:[12,2,1,""],release:[12,2,1,""],type:[12,3,1,""]},"autopilot.hardware.gpio.Digital_Out":{_series_script:[12,2,1,""],delete_script:[12,2,1,""],output:[12,3,1,""],pigs_function:[12,3,1,""],pulse:[12,2,1,""],release:[12,2,1,""],series:[12,2,1,""],set:[12,2,1,""],stop_script:[12,2,1,""],store_series:[12,2,1,""],toggle:[12,2,1,""],turn:[12,2,1,""],type:[12,3,1,""]},"autopilot.hardware.gpio.GPIO":{init_pigpio:[12,2,1,""],pin:[12,2,1,""],polarity:[12,2,1,""],pull:[12,2,1,""],release:[12,2,1,""],trigger:[12,2,1,""]},"autopilot.hardware.gpio.LED_RGB":{_series_script:[12,2,1,""],flash:[12,2,1,""],output:[12,3,1,""],pin:[12,2,1,""],pin_bcm:[12,2,1,""],pull:[12,2,1,""],pulse:[12,2,1,""],range:[12,2,1,""],release:[12,2,1,""],set:[12,2,1,""],toggle:[12,2,1,""],type:[12,3,1,""]},"autopilot.hardware.gpio.PWM":{output:[12,3,1,""],pigs_function:[12,3,1,""],polarity:[12,2,1,""],range:[12,2,1,""],release:[12,2,1,""],set:[12,2,1,""],type:[12,3,1,""]},"autopilot.hardware.gpio.Solenoid":{DURATION_MIN:[12,3,1,""],dur_from_vol:[12,2,1,""],duration:[12,2,1,""],open:[12,2,1,""],output:[12,3,1,""],type:[12,3,1,""]},"autopilot.hardware.i2c":{I2C_9DOF:[13,1,1,""],MLX90640:[13,1,1,""]},"autopilot.hardware.i2c.I2C_9DOF":{ACCELRANGE_16G:[13,3,1,""],ACCELRANGE_2G:[13,3,1,""],ACCELRANGE_4G:[13,3,1,""],ACCELRANGE_8G:[13,3,1,""],GYROSCALE_2000DPS:[13,3,1,""],GYROSCALE_245DPS:[13,3,1,""],GYROSCALE_500DPS:[13,3,1,""],MAGGAIN_12GAUSS:[13,3,1,""],MAGGAIN_16GAUSS:[13,3,1,""],MAGGAIN_4GAUSS:[13,3,1,""],MAGGAIN_8GAUSS:[13,3,1,""],accel_range:[13,2,1,""],acceleration:[13,2,1,""],gyro:[13,2,1,""],gyro_scale:[13,2,1,""],mag_gain:[13,2,1,""],magnetic:[13,2,1,""],temperature:[13,2,1,""]},"autopilot.hardware.i2c.MLX90640":{ALLOWED_FPS:[13,3,1,""],SHAPE_SENSOR:[13,3,1,""],_grab:[13,2,1,""],_threaded_capture:[13,2,1,""],_timestamp:[13,2,1,""],capture_init:[13,2,1,""],fps:[13,2,1,""],init_cam:[13,2,1,""],integrate_frames:[13,2,1,""],interpolate:[13,2,1,""],interpolate_frame:[13,2,1,""],release:[13,2,1,""],type:[13,3,1,""]},"autopilot.hardware.usb":{Scale:[14,1,1,""],Wheel:[14,1,1,""]},"autopilot.hardware.usb.Scale":{MODEL:[14,3,1,""]},"autopilot.hardware.usb.Wheel":{MODES:[14,3,1,""],MOVE_DTYPE:[14,3,1,""],THRESH_TYPES:[14,3,1,""],assign_cb:[14,2,1,""],calc_move:[14,2,1,""],check_thresh:[14,2,1,""],input:[14,3,1,""],l_clear:[14,2,1,""],l_measure:[14,2,1,""],l_stop:[14,2,1,""],release:[14,2,1,""],start:[14,2,1,""],thresh_trig:[14,2,1,""],trigger:[14,3,1,""],type:[14,3,1,""]},"autopilot.prefs":{add:[15,4,1,""],compute_calibration:[15,4,1,""],git_version:[15,4,1,""],init:[15,4,1,""],prefdict:[15,5,1,""]},"autopilot.setup":{setup_pilot:[17,0,0,"-"],setup_scale:[18,0,0,"-"],setup_terminal:[19,0,0,"-"]},"autopilot.setup.setup_pilot":{HardwareForm:[17,1,1,""],PilotSetupForm:[17,1,1,""],SetupApp:[17,1,1,""],make_dir:[17,4,1,""],unfold_values:[17,4,1,""]},"autopilot.setup.setup_pilot.HardwareForm":{afterEditing:[17,2,1,""],create:[17,2,1,""]},"autopilot.setup.setup_pilot.PilotSetupForm":{afterEditing:[17,2,1,""],create:[17,2,1,""]},"autopilot.setup.setup_pilot.SetupApp":{onStart:[17,2,1,""]},"autopilot.setup.setup_terminal":{SetupApp:[19,1,1,""],TerminalSetupForm:[19,1,1,""],make_dir:[19,4,1,""],unfold_values:[19,4,1,""]},"autopilot.setup.setup_terminal.SetupApp":{onStart:[19,2,1,""]},"autopilot.setup.setup_terminal.TerminalSetupForm":{afterEditing:[19,2,1,""],create:[19,2,1,""]},"autopilot.stim":{managers:[21,0,0,"-"],sound:[22,0,0,"-"]},"autopilot.stim.managers":{Bias_Correction:[21,1,1,""],Proportional:[21,1,1,""],Stim_Manager:[21,1,1,""],init_manager:[21,4,1,""]},"autopilot.stim.managers.Bias_Correction":{next_bias:[21,2,1,""],thresholded_linear:[21,2,1,""],update:[21,2,1,""]},"autopilot.stim.managers.Proportional":{init_sounds_grouped:[21,2,1,""],init_sounds_individual:[21,2,1,""],next_stim:[21,2,1,""],set_triggers:[21,2,1,""],store_groups:[21,2,1,""]},"autopilot.stim.managers.Stim_Manager":{compute_correction:[21,2,1,""],do_bias:[21,2,1,""],do_correction:[21,2,1,""],end:[21,2,1,""],init_sounds:[21,2,1,""],make_punishment:[21,2,1,""],next_stim:[21,2,1,""],play_punishment:[21,2,1,""],set_triggers:[21,2,1,""],update:[21,2,1,""]},"autopilot.stim.sound":{jackclient:[23,0,0,"-"],pyoserver:[24,0,0,"-"],sounds:[25,0,0,"-"]},"autopilot.stim.sound.jackclient":{BLOCKSIZE:[23,5,1,""],CONTINUOUS:[23,5,1,""],CONTINUOUS_LOOP:[23,5,1,""],CONTINUOUS_QUEUE:[23,5,1,""],FS:[23,5,1,""],JackClient:[23,1,1,""],PLAY:[23,5,1,""],QUEUE:[23,5,1,""],Q_LOCK:[23,5,1,""],SERVER:[23,5,1,""],STOP:[23,5,1,""]},"autopilot.stim.sound.jackclient.JackClient":{boot_server:[23,2,1,""],process:[23,2,1,""],quit:[23,2,1,""],run:[23,2,1,""]},"autopilot.stim.sound.pyoserver":{pyo_server:[24,4,1,""]},"autopilot.stim.sound.sounds":{File:[25,1,1,""],Gap:[25,1,1,""],Noise:[25,1,1,""],SOUND_LIST:[25,5,1,""],STRING_PARAMS:[25,5,1,""],Speech:[25,1,1,""],Tone:[25,1,1,""],int_to_float:[25,4,1,""]},"autopilot.stim.sound.sounds.File":{PARAMS:[25,3,1,""],init_sound:[25,2,1,""],type:[25,3,1,""]},"autopilot.stim.sound.sounds.Gap":{PARAMS:[25,3,1,""],buffer:[25,2,1,""],chunk:[25,2,1,""],init_sound:[25,2,1,""],play:[25,2,1,""],type:[25,3,1,""]},"autopilot.stim.sound.sounds.Noise":{PARAMS:[25,3,1,""],init_sound:[25,2,1,""],type:[25,3,1,""]},"autopilot.stim.sound.sounds.Speech":{PARAMS:[25,3,1,""],type:[25,3,1,""]},"autopilot.stim.sound.sounds.Tone":{PARAMS:[25,3,1,""],init_sound:[25,2,1,""],type:[25,3,1,""]},"autopilot.tasks":{CHILDREN_LIST:[26,5,1,""],TASK_LIST:[26,5,1,""],children:[27,0,0,"-"],free_water:[28,0,0,"-"],graduation:[29,0,0,"-"],nafc:[30,0,0,"-"],task:[31,0,0,"-"]},"autopilot.tasks.children":{Transformer:[27,1,1,""],Video_Child:[27,1,1,""],Wheel_Child:[27,1,1,""]},"autopilot.tasks.children.Transformer":{l_process:[27,2,1,""],noop:[27,2,1,""]},"autopilot.tasks.children.Video_Child":{PARAMS:[27,3,1,""],noop:[27,2,1,""],start:[27,2,1,""],stop:[27,2,1,""]},"autopilot.tasks.children.Wheel_Child":{HARDWARE:[27,3,1,""],PARAMS:[27,3,1,""],STAGE_NAMES:[27,3,1,""],end:[27,2,1,""],noop:[27,2,1,""]},"autopilot.tasks.free_water":{Free_Water:[28,1,1,""]},"autopilot.tasks.free_water.Free_Water":{DATA:[28,3,1,""],HARDWARE:[28,3,1,""],PARAMS:[28,3,1,""],PLOT:[28,3,1,""],STAGE_NAMES:[28,3,1,""],TrialData:[28,1,1,""],end:[28,2,1,""],response:[28,2,1,""],water:[28,2,1,""]},"autopilot.tasks.free_water.Free_Water.TrialData":{columns:[28,3,1,""]},"autopilot.tasks.graduation":{Accuracy:[29,1,1,""],GRAD_LIST:[29,5,1,""],Graduation:[29,1,1,""],NTrials:[29,1,1,""]},"autopilot.tasks.graduation.Accuracy":{COLS:[29,3,1,""],PARAMS:[29,3,1,""],update:[29,2,1,""]},"autopilot.tasks.graduation.Graduation":{COLS:[29,3,1,""],PARAMS:[29,3,1,""],update:[29,2,1,""]},"autopilot.tasks.graduation.NTrials":{PARAMS:[29,3,1,""],update:[29,2,1,""]},"autopilot.tasks.nafc":{Nafc:[30,1,1,""],Nafc_Gap:[30,1,1,""]},"autopilot.tasks.nafc.Nafc":{HARDWARE:[30,3,1,""],PARAMS:[30,3,1,""],PLOT:[30,3,1,""],STAGE_NAMES:[30,3,1,""],TrialData:[30,1,1,""],discrim:[30,2,1,""],flash_leds:[30,2,1,""],punish:[30,2,1,""],reinforcement:[30,2,1,""],request:[30,2,1,""],respond:[30,2,1,""],stim_end:[30,2,1,""],stim_start:[30,2,1,""]},"autopilot.tasks.nafc.Nafc.TrialData":{columns:[30,3,1,""]},"autopilot.tasks.nafc.Nafc_Gap":{PARAMS:[30,3,1,""],end:[30,2,1,""]},"autopilot.tasks.task":{Task:[31,1,1,""]},"autopilot.tasks.task.Task":{HARDWARE:[31,3,1,""],PARAMS:[31,3,1,""],PLOT:[31,3,1,""],STAGE_NAMES:[31,3,1,""],TrialData:[31,3,1,""],end:[31,2,1,""],flash_leds:[31,2,1,""],handle_trigger:[31,2,1,""],init_hardware:[31,2,1,""],set_leds:[31,2,1,""],set_reward:[31,2,1,""]},"autopilot.viz":{psychometric:[33,0,0,"-"],trial_viewer:[34,0,0,"-"]},"autopilot.viz.psychometric":{calc_psychometric:[33,4,1,""],plot_psychometric:[33,4,1,""]},"autopilot.viz.trial_viewer":{load_subject_data:[34,4,1,""],load_subject_dir:[34,4,1,""],step_viewer:[34,4,1,""],trial_viewer:[34,4,1,""]},autopilot:{core:[0,0,0,"-"],hardware:[10,0,0,"-"],prefs:[15,0,0,"-"],setup:[16,0,0,"-"],stim:[20,0,0,"-"],tasks:[26,0,0,"-"],viz:[32,0,0,"-"]}},objnames:{"0":["py","module","Python module"],"1":["py","class","Python class"],"2":["py","method","Python method"],"3":["py","attribute","Python attribute"],"4":["py","function","Python function"],"5":["py","data","Python data"]},objtypes:{"0":"py:module","1":"py:class","2":"py:method","3":"py:attribute","4":"py:function","5":"py:data"},terms:{"01t17":6,"100u":12,"1mhz":[13,37],"2000m":40,"20m":44,"2afc":[8,26,44],"2afc_gap":26,"30fp":11,"30khz":46,"32x24px":37,"3rd":[35,46],"400khz":13,"50m":11,"64fp":[9,13,37],"807693v1":[36,42],"8mb":40,"9dof":[13,37],"abstract":[38,46],"boolean":43,"break":[2,44],"byte":12,"case":[8,11,12,23,37,38,46],"class":[2,3,6,8,10,11,12,13,14,17,19,21,23,25,26,27,28,29,30,31,37,39,42,44,45],"default":[2,3,6,8,11,12,13,17,19,23,37,38,40,41,43,44,46],"enum":46,"export":[6,37],"final":[25,37,43,44],"float":[3,6,8,11,12,13,21,25,29,30,31,38,43],"function":[1,2,4,7,8,10,11,12,14,15,17,19,21,24,25,31,33,34,37,38,40,43,46],"import":[12,15,37,38,42,43,46],"int":[2,3,6,8,10,11,12,13,14,21,23,25,27,28,29,30,31,38,43],"long":[3,6,10,46],"new":[2,3,6,11,21,35,43,44,46],"return":[2,3,6,8,9,10,11,12,13,14,15,21,24,25,27,28,29,30,31,33,34,37,43],"short":[42,43,46],"static":40,"super":43,"switch":[37,39,44,46],"true":[2,3,6,10,11,12,14,21,24,25,27,29,30,31,34,38,43,44,46],"try":[2,8,11,38,43],"while":[6,11,30,37,40,43,46],Added:37,Adding:44,And:[15,40,43],Are:[12,21,40],FPS:[11,13],For:[11,12,21,35,38,42,44,45,46],IDs:12,Its:[38,45],NOT:[15,23],Not:[2,3,8,10,11,14,18,21,38],One:[11,21],Pis:[42,44,45],That:[37,44,46],The:[2,3,6,8,10,11,12,13,17,21,23,25,27,28,37,38,39,40,41,42,44,46],Then:[2,43],There:[2,38,43,46],These:[2,12,25,39,40,43,44],Use:[11,24,31,40,46],Used:[2,3,6,12,43],Uses:[12,13,14],Using:11,Was:21,With:[30,38,46],Yes:[6,44],__contains__:[2,8],__del__:[10,38],__delitem__:[2,8],__dict__:2,__getitem__:[2,8],__init__:[8,10,38,43,46],__setitem__:[2,8],_cam:11,_camera_method:11,_captur:11,_clean_valu:12,_deserialize_numpi:37,_frame:13,_grab:[11,13],_grab_ev:13,_mous:43,_output_filenam:11,_process:11,_record:43,_serialize_numpi:[2,37],_series_script:12,_step_nam:6,_stream:2,_threaded_captur:13,_timestamp:[11,13],_write_deinit:11,_write_fram:11,_write_q:11,abbrevi:43,abil:[6,37,46],abl:[10,46],about:[6,11,25,37,46],abov:[14,21,23,29,40,43,44],absolut:[11,37],abyz:12,accel:13,accel_rang:13,acceler:[13,46],acceleromet:[13,37,46],accelrange_16g:13,accelrange_2g:13,accelrange_4g:13,accelrange_8g:13,accept:[10,11,12,14],access:[6,23,37,40,43,44],accord:[2,3,6,11,13,21,25,37,42],accordingli:[2,12,13,38,46],accur:[37,43,46],accuraci:[29,42,44],achiev:37,acquir:[11,37,46],acquisit:[11,13,46],acquisition_mod:11,acquisitionframer:11,across:[13,37,40,43,46],activ:[8,10,11,14,23,30,43],actual:[38,42,44],adafruit:13,adapt:46,add:[6,12,15,21,38,40,43,44,46],added:[2,25,37,40,44,46],adding:[12,44],addit:[10,12,37,40,42,46],addition:[43,46],address:[2,3,17,40,43,44],addus:40,adir:[17,19],admin:40,adopt:46,advanc:[40,43],advantag:11,advis:44,aesthet:37,affect:38,after:[2,6,8,11,12,15,23,29,34,37,38,40,41,43,44],afteredit:[17,19],afterward:28,agent:[15,27,37,39,41,42,46],alert:[2,28,46],algorithm:46,align:[43,46],aliv:[2,3,8,12,23],all:[2,6,10,11,12,19,21,25,28,29,31,37,38,39,40,41,43,44,46],allcap:15,alloc:[11,40],allow:[11,12,28,37,38,40,42,43,44,45,46],allow_repeat:[28,44],allowed_fp:13,almost:45,along:[2,6,25,43],alongsid:[31,46],alreadi:[2,6,11,12,40,46],alsa:40,alsafil:40,also:[8,10,12,37,38,40,43,45,46],altair:[33,37],altern:[6,30,44],alwai:[2,8,10,11,13,37,40,46],ambit:37,amount:[3,12,23],amp:40,amplifi:40,amplitud:[25,30,44],analog:42,analysi:46,analyz:46,ancient:46,angl:43,ani:[2,3,8,10,11,12,13,14,15,17,19,28,29,31,38,39,40,43,44,45,46],anim:44,anod:[12,38],anoth:[2,11,12,37,38,43,46],answer:[21,30,44],anyth:[6,8,10],anywai:[6,11],api:[38,42,45,46],aplai:40,apnpsapplicationmanag:[17,19],apparatu:42,appear:44,append:[6,11,38],appli:[12,38,46],applic:[11,38,46],apply_along:6,appropri:[2,6,15,37,43,44],apt:40,arab:40,arbitrari:[21,42,45,46],architectur:[37,46],aren:37,arena:[37,44,46],arg:[2,8,11,13,17,19,21,28,30,31,43],argument:[8,10,11,12,28,43],aris:6,arm:40,around:[11,12,37,43,46],arrai:[2,13,14,23,25,37,38,43,45,46],arrang:37,asctim:[2,3],ask:[2,3,17,43],asound:40,assembl:46,assign:[6,12,31,43,44],assign_cb:[10,12,14,38],assign_protocol:6,assum:[12,38,43,46],astunpars:40,asynchron:30,attempt:[2,8,11,37,38,44],attr:[11,12,13],attr_typ:11,attr_type_nam:11,attribut:[2,6,8,10,11,12,13,14,25,27,28,29,30,31,37,38],attributeerror:43,audibl:40,audio:[3,17,23,24,25,37,39,46],audioserv:[3,15,17,25],auditori:46,auto:[11,46],autom:[41,42,46],automat:[11,37,44,45,46],autopilot:[2,3,5,9,11,12,13,14,15,17,19,21,25,26,27,28,29,30,36,37,38,39,40,41,42,43,44,45],autopilot_root:15,autpilot:35,avail:[11,12,37,40,43,46],averag:[11,13,37],avoid:[11,25,37,40,46],await:13,awar:44,awkward:[37,46],axi:[13,43],b64:2,b64encod:2,back:[2,3,31,43],backend:[11,40],background:[12,23,38],bad:[6,10,43],bail:[30,31,43],balanc:37,bar:[31,34,43,46],base64:2,base:[2,3,6,8,10,11,12,13,14,17,19,21,23,25,27,28,29,30,31,37,38,40,41,44,46],base_dir:[40,41],base_path:11,basedir:[3,15,17,19,41,44],baselin:6,baseprocess:11,basestr:6,bash:[10,46],basic:[21,37,38,40,41,43,44,46],batch:[37,46],baudrat:[9,13,37],bcm:[10,12,31,38],bcm_to_board:10,beam:44,beambreak:[3,17,31,40,43],beauti:46,becaus:[11,37,43,44],becom:46,been:[2,8,11,12,25,37,38,39,40,43,45,46],befor:[6,8,11,15,23,30,37,38,40,43,46],begin:[11,37,43,44],beginacquisit:11,begun:37,behav:[37,43],behavior:[2,42,45],behind:2,beholden:46,being:[2,6,11,12,37,38,43,46],bell:38,below:[14,41,43,44],better:[25,37,40,46],between:[2,3,10,21,26,28,29,37,39,40,42,43,46],bia:[21,30,43],bias:21,bias_correct:[21,30],bias_mod:[30,43,44],bias_threshold:[30,43],bidi:40,bin:[11,15,46],binari:[3,11,17,38],bind:[2,40],biograph:[6,44],biographi:6,biography_tab:6,biorxiv:[36,42],bit:[13,25,37,38,43,46],bittorr:46,blank:[2,43],blank_l:3,blend:46,blink:[12,38],block:[3,8,11,38,43],blocksiz:23,blosc:[2,11,37,40,41],blue:[12,30,43],bluealsa:40,bluetooth:40,board:[3,10,12,17,37,38,45],board_to_bcm:[10,38],bodi:2,bom:46,bool:[2,6,8,10,11,12,14,21,24,25,28,29,30,31,34,38,43],boolcol:43,boost:37,boot:[3,13,17,23,24,31,40],boot_serv:23,border:46,both:[12,38,40,43,46],both_edg:12,bother:25,bottleneck:46,bound:2,brand:11,brandi:46,breviti:38,brew:41,bridg:10,bright:46,bring:12,broadcom:[10,12],brought:[37,46],brows:46,buffer:[11,13,23,25,37,40,43],bug:37,bugfix:35,build:[37,40,43,44,46],build_pi:46,built:[21,38,42,46],bunch:37,button:[44,46],c3c7dcb32f164e2740b4fca98237908e74a00a4c:15,c_doubl:11,cach:[3,23,43],cad:46,cal_progress:3,calc_mov:14,calc_psychometr:33,calcul:[14,21,33,37,43],calibr:[3,12,13,15,25,37,41,46],calibrate_port:[3,12],calibration_curv:3,call:[2,3,6,8,10,11,12,13,14,17,19,21,23,25,30,31,33,37,38,39,40,43,44,46],callabl:[8,12,21,28,43],callback:[9,10,12,21,30,31,38,43,46],callback_fn:[12,38],cam:[11,27],cam_list:11,camera:[10,13,27,37,45,46],camera_cv:[11,37],camera_idx:11,camera_spinnak:[11,37],cameralist:11,can:[2,6,8,10,11,12,13,21,23,25,30,36,37,38,40,42,43,44,45,46],cancel:[12,43],cannot:12,canon:12,cap_prop_fp:11,cap_prop_frame_width:11,cap_prop_pos_msec:11,capabl:[9,13,37,46],captur:[9,11,13,37,43,45],capture_:11,capture_deinit:11,capture_directori:11,capture_init:[11,13],card:40,care:[10,38,40],carri:28,cascad:46,categori:44,cathod:[12,38],caus:[6,8,23,40,46],cautiou:11,ceaseless:37,center:[11,30,43,44],certain:43,cffi:40,cfg:40,challeng:44,chanc:21,chance_bar:[30,31,43],chance_level:43,chang:[2,3,6,12,25,27,36,37,40,41,43,44,46],change_to_blu:43,changelog:45,chann:12,channel:[3,11,12,17,23,40],charact:[30,43],chart:33,check:[2,3,6,11,14,37,40,43,45,46],check_thresh:[14,43],checkbox:43,child:[2,26,27,37,40,46],childid:43,children:[2,26,42,43,46],children_list:[26,43],chip:10,choic:[21,30,44],choos:[21,40,43,46],chosen:46,chunk:25,circuitfun:13,circular:43,clariti:38,classic:46,clean:[13,38],cleanli:46,cleanup:38,clear:[3,12,14,23,38,43],clear_cb:12,click:[38,44,46],client:[10,23,40],clock:[40,46],clone:40,close:[6,12,37],close_hdf:6,clumsi:46,cmake:40,coars:6,code:[13,35,38,43,46],coeffici:12,coerc:8,coerce_discret:8,coher:[2,3,8],col:[8,29,38,43],col_nam:6,collabor:46,collect:[21,27,31,34,37,38,43,44,46],collis:46,color:[3,12,17,31,38,40,43],color_dict:31,color_seri:38,column:[6,8,25,28,29,30,33,37,43,44,46],com:[8,11,12,13,14,35,37,40],combin:[12,13,25,37,42,45],come:[25,37,43,44],command:[10,17,19,40,44,46],comment:40,commit:[6,15,35],common:[12,37,38,44,46],commonli:[12,43],commun:[2,3,38,42,45,46],compat:[6,37],compil:[11,40],complet:[11,12,30,31,37,42,43],complex:[2,43,45,46],compon:[37,42,45],comprehens:42,compres:11,compress:[2,11,37,40,45,46],comput:[2,3,12,14,21,30,37,38,42,43,46],compute_calibr:15,compute_correct:21,concept:43,concurr:46,condit:[27,46],conf:40,config:[13,15,37,40],configobj:40,configur:[3,11,17,37,38,40,41,42,43,44,46],confin:43,confirm:[2,37],conflict:37,confus:[37,45],confusingli:40,connect:[2,3,12,13,17,23,37,38,40,43,46],consid:[29,46],consist:[36,42,44],consol:46,conson:25,constant:38,constitut:42,construct:[8,12,42,43],constructor:8,consum:[11,44,46],contain:[2,28,37,46],content:[36,42],context:[2,11,23],continu:[2,3,11,13,14,21,23,25,30,31,37,43,46],continuous_cycl:23,continuous_data:6,continuous_loop:23,continuous_queu:23,continuousdata:[43,46],contribut:43,control:[2,3,11,17,37,40,41,42,43,44,46],conveni:[12,38],convers:11,convert:[10,11,12,25,37,38,46],cookbook:46,cool_new_step:44,coordin:[3,42,45,46],copi:[2,40],core:[2,37,43,45],correct:[21,28,29,30,31,37,40,43,44],correction_pct:[21,30,43,44],correction_tri:[21,30,43,46],correctli:[9,44],correspond:[12,29,40,44,46],corrupt:6,costli:46,could:[37,38,40,46],count:[2,12,28,29,30,43],counter:[2,29],cpu:40,cpufreq:40,creat:[2,3,6,11,12,17,19,25,37,38,39,40,41,43,45,46],creation:[2,46],credenti:40,credit:40,crew:46,criteria:[29,42,44],critic:[37,40],cross:43,crude:46,css:37,csv:[6,11],ctl:[11,40],ctrl:40,ctx:2,cumbersom:46,current:[2,3,6,8,10,11,12,15,17,21,28,30,37,38,39,40,41,42,43,44,46],current_stag:[30,43],current_tri:[6,28,29,30,43,44],curv:[3,33,37],custom:[12,25,40],customiz:42,cute:12,cv2:11,cycl:[11,12,23,28,37,38,43,46],cycle_widget:17,dac:40,dacplu:40,daemon:[2,10,12,37,38,43,44],dag:46,dai:3,dalsa:[3,17,40],data:[2,3,6,10,11,12,15,19,23,25,26,28,29,30,31,33,34,37,40,41,42,44,45,46],data_dir:34,data_queu:6,data_thread:6,databas:46,datadir:[3,6,15,19,41],datafram:[6,8,33],date:6,date_protocol_nam:6,datetim:[2,13,30,43],dc_timestamp:[30,31,43],deadlock:8,dealer:[2,46],debug:[11,15,24,37,43],decid:[8,21,37],decim:8,decis:37,declar:37,decod:[2,46],decompress:2,decor:46,decreas:40,decrement:2,dedic:37,def:[38,43],defin:[2,3,6,10,17,25,29,31,38,42,43],degre:13,delai:[38,43],delayed_set:43,delet:[6,12,38],delete_script:12,delin:43,deliv:[30,43,45],deliveri:[12,37,43,45],demand:[10,40],demonstr:43,denot:43,dep:40,depend:[3,10,11,14,21,25,30,37,40,41,43,46],deprec:40,depth:[37,40,43],dequ:21,deriv:[17,19],descend:46,describ:[6,10,21,37,43,44,46],descript:[2,6,11,12,26,28,30,31,37,38,42,43,46],descriptor:[11,37,43],deseri:37,design:[10,27,37,38,43,45,46],desir:[3,11,12],desktop:42,destroi:37,detail:[6,11,35,37,42,43,45],detect:[12,38,44],determin:[2,23,46],dev:40,dev_debug:11,develop:[10,37,39,40,41,42,43,45,46],devic:[2,10,11,12,13,14,37,38,40,43,46],device_info:11,deviceinform:11,dflt:[6,28,30],dhw:[3,17,40],dialog:37,dict:[2,3,6,8,10,11,12,15,17,19,21,27,28,30,31,38,43],dictionari:[2,3,6,8,10,11,15,21,27,28,31,37,40,43,44],did:[29,46],did_gradu:6,didn:[6,37],die:2,differ:[2,11,13,21,29,37,38,39,42,43,46],differenti:37,difficult:46,difficulti:11,digi_in:12,digi_out:[14,43],digit:[12,14,18,37,38,40],digital_in:[12,28,30,37,38,40],digital_out:[12,14,27,37,43],dir:[6,14],direct:[2,12,21,38,40,46],directli:[2,11,23,40,43,46],directori:[3,11,15,17,19,34,40,41,44,46],directory_writ:11,disabl:[40,44],discard:23,discourag:24,discret:[8,10,37],discrim:30,discrim_plai:[30,43],discrimin:44,discriminand:[30,33],discuss:[37,38,43,45],dish:46,disk2:40,disk:[11,40,43],diskutil:40,disorgan:37,dispens:[28,40],dispers:3,displai:[37,44,46],dissect:38,dist:[14,43],distanc:[14,27,43],distinct:11,distinctli:46,distractor:[21,30,43],distribut:[2,42,45,46],disturb:46,distutil:40,divid:[12,38],dlc:46,dlc_latenc:26,dmesg:11,do_bia:[21,43],do_correct:[21,43],do_log:37,do_return:15,do_trigg:43,dob:6,doc:[3,5,6,10,12,15,17,37,46],docker:46,document:[11,38,40,42,43,45],doe300:46,doe:[2,6,11,12,25,39,40,41,43,44],doesn:[11,12,17,19,43,44,46],doing:[8,21,25,29],don:[2,37,38,40,42,43,46],done:46,dos:46,doubl:[2,46],down:[3,12,17,28,37,38,43],download:[40,46],download_box:37,dpkg:40,draft:21,drag:46,dramat:[37,41],draw:[31,43],draw_line_at:19,drawfp:[2,15,19,37,41],drawn:43,drift:43,drive:3,driver:[9,37,40],drop:[2,23,46],dropout:24,drown:37,dtoverlai:40,dtparam:[13,40],dtype:38,due:24,dump:[2,6,11,23,37,40],duplic:[37,46],dur_from_vol:12,durat:[3,11,12,21,25,28,30,31,38,43,44],duration_min:12,dure:[11,21,31],duti:[12,38],dutycycl:12,each:[2,3,6,10,11,12,17,23,25,27,30,31,33,34,37,38,39,41,42,43,44,46],earli:43,earlier:44,easier:[10,37,39,45,46],easili:37,echo:[11,40],edg:[12,38],edit:[13,40,44,46],effect:[23,25,42,43],effici:[45,46],effort:[37,46],eft:[43,44],egregi:43,either:[2,3,6,8,11,12,27,37,38,40,43,44,46],either_edg:12,elaps:43,element:[43,46],els:[6,8,12,25,29,43],elsewher:25,elud:46,embarassingli:37,empti:[11,23,38],enabl:[11,12,37,40,44,46],encapsul:2,encod:[2,11,37],end:[6,11,12,21,23,27,28,30,31,37,43,44],end_thread:38,endless:43,enforc:[23,38],engin:37,enhanc:11,enough:37,ensur:[6,43,46],ensure_structur:6,enter:43,entir:[40,46],entri:[2,3,40,43,44],entropi:37,environ:[39,46],equal:38,equival:6,error:[8,11,37,45,46],especi:38,esprima:40,essenti:[39,40],establish:[3,38],etc:[2,6,10,11,12,37,40,43,44,46],eter:46,even:[10,37],event:[2,3,6,10,11,12,13,23,28,30,31,37,38,43,45,46],everi:[2,3,6,10,13,14,37,44,45,46],everyon:2,everyth:[31,39,46],evil:43,ewm:34,exactli:45,exampl:[15,25,37,38,40,43,44,46],except:[8,11,37,38,43,44],excess:38,execut:[11,12],exist:[2,6,9,11,17,19,36,43,44,46],exit:[12,40],expand:[2,46],expand_arrai:2,expand_on_rec:2,expect:[11,37,38,43,46],experi:[40,42,45,46],experiment:[40,42,46],explain:43,explan:36,explicit:[6,37],explicitli:[2,10,11,12,37,38,43],exposur:11,exposuretim:11,extend:[40,42,43,46],extens:[8,11,43,46],extern:[2,11,35,45,46],extra:[38,46],extract:[37,38,43,44],extran:25,extraordinarili:11,extrem:38,face:[2,41,42],fail:37,fakekm:40,fall:[12,38,43,46],falling_edg:12,fals:[2,6,8,10,11,12,14,15,17,21,24,25,28,29,30,31,34,38,43,44],far:46,faster:[2,14,40],fastest:40,favor:21,fbdev:40,fdisk:40,featur:[11,35,46],feed:3,feel:46,few:[10,37,38,40,41,43,44],ffmpegwrit:11,field:[2,14,30,44,46],file:[2,3,6,10,11,15,17,25,34,37,39,40,43,44,45,46],file_block:[2,3],file_db:8,file_nam:8,filehandl:[2,3],filenam:[6,11],filenod:6,filestructur:6,filesystem:40,fill:[23,44],filter:[25,38,46],final_recipi:37,find:2,find_packag:46,finish:[3,11,30,43],finish_off:12,finit:[11,42,46],firmwar:11,first:[21,23,28,30,37,43,44,46],first_param:27,fit:[12,43],fix:[6,37,40,43,46],fixat:43,flag:[2,3,6,23,37,40,43],flash:[12,30,31,37,38,43],flash_block:38,flash_l:[30,31,43],flatten:[6,46],flexibl:[10,37,40],flicker:46,flir:[11,37],float32col:[6,43],float64col:43,flush:6,flush_curr:6,fmform:[17,19],follow:[3,6,11,17,19,21,40,43,46],forc:[30,44,46],force_to:2,forev:40,form:[8,11,12,17,46],formal:[27,38,39,46],format:[2,3,6,21,37,43,44,46],formatt:[2,3,46],formdefault:17,forth:46,forthcom:42,forum:37,forward:2,found:[11,37,42,44],fps:[11,13,19,41],fraction:8,frame:[2,11,13,17,23,37,40,45,46],frame_trigg:11,framer:[11,13],framework:[43,45,46],free:[26,44],free_wat:26,freeli:[44,46],freetyp:40,freez:46,freq:43,frequenc:[12,21,25,37,38,43,44],frequent:37,from:[2,3,6,8,10,11,12,13,15,17,19,21,23,28,29,37,38,40,41,42,43,44,45,46],fromstr:11,full:[6,9,11,12,34,35,37,46],fuller:46,further:[2,11,37,43,46],futur:[2,3,6,8,21,37,38,39,40,43,44,45,46],gabor:43,gain:13,gap:[25,37],gap_zero:25,garbag:38,gather:43,gauss:13,gave:43,gdc:12,gener:[2,6,10,11,12,15,19,21,23,24,25,31,37,38,41,42,43,44,46],genicam:11,genotyp:44,get:[2,3,6,10,11,13,15,29,31,37,38,40,43,46],get_current_tim:37,get_ip:[2,3],get_nam:10,get_step_histori:6,get_stream:[2,11,37,45],get_sync_tim:37,get_thi:37,get_timestamp:[2,6],get_trial_data:6,get_weight:6,getattr:43,getter:[37,46],getvalu:11,gevent:40,ghasp:46,git:[6,15,19,37,40,41,46],git_vers:15,github:[12,35,37,40,46],gitlab:40,gitpython:40,give:[2,3,21,30,37,38,40,43,44,46],given:[2,3,6,8,10,11,12,14,21,25,29,37,38,40,43],glass:46,glfw:40,glitch:40,global:[15,46],gnueabihf:40,goal:[37,39,46],god:46,going:43,gonna:43,gonogo:[26,43],good:[37,38,40,43,46],googl:37,got:[37,40],govern:39,governor:40,gpio:[3,9,10,17,27,28,30,31,37,40,43,45,46],gpu:40,grab:[11,40],grad:34,grad_data:34,grad_list:29,graduat:[6,26,34,42,44],gram:6,grammar:46,graphic:[40,43,46],grate:43,green:[12,43],griddata:13,ground:13,group:[3,6,8,21,37,40],group__timefunct:46,group_nam:21,guarante:[12,46],guess:10,gui:[0,2,5,19,37,41,42,43,46],guid:[38,45],gyro:13,gyro_scal:13,gyroscale_2000dp:13,gyroscale_245dp:13,gyroscale_500dp:13,gyroscop:[13,37],h5f:6,had:[37,43,46],half:38,hand:38,handl:[2,3,10,11,12,23,37,38,43,46],handle_listen:2,handle_trigg:[10,14,31,43,46],handler:[2,3,46],handshak:[2,3,44],happen:[8,37,46],hard:[37,46],hardcod:[15,37,43,44],hardwar:[2,3,9,11,12,13,14,17,27,28,30,31,37,40,42,44,45,46],hardwareform:17,harshli:46,has:[2,6,8,10,11,14,21,25,28,37,38,40,43,44,45,46],hash:[6,15,46],hash_tabl:6,hasn:11,have:[2,6,10,11,12,21,27,28,29,30,37,38,39,40,41,42,43,44,45,46],haven:[2,37],hciuart:40,hdf5:[6,34,41,44,46],hdf:6,header:40,headless:40,heartbeat:46,heavili:42,height:[11,43],held:[11,38],help:[17,38,40],here:[2,6,11,12,15,17,19,29,38,40,43,44],hidden:46,hierarch:46,hifi:40,hifiberri:[40,46],high:[3,12,31,37,38,40,44],higher:40,highli:[11,37,43],histori:[6,21,43,44],history_t:6,hit:2,hold:[2,11,15],hole:44,home:11,honor:46,hope:37,horizont:[11,43],hover:46,how:[2,3,11,12,21,31,38,40,43,44,46],howev:[6,12,44],howto:46,html:[5,12,46],http:[5,8,11,12,13,35,36,37,40,42],human:[11,43,46],humbl:25,i2c1:40,i2c:[9,10,12,37,40],i2c_9dof:[13,37,45],i2c_arm:40,i2c_arm_baudr:13,i2s:40,i32:28,ideal:[12,40,46],ident:[2,43,44],identifi:[2,3,11,40,43],idl:[3,11],ight:[43,44],ignor:[3,6,11,29,40],iidc:11,iis:11,illustr:38,imag:[11,40,46],image1:11,image_path:11,imageitem:37,imageitem_timedupd:37,img:40,img_opt:11,immatur:[43,46],immedi:43,implement:[2,3,6,8,10,14,18,21,25,29,38,43,46],implicit:46,implicitli:[37,44],imposs:46,improv:[37,39],incap:46,includ:[2,3,6,9,11,12,37,43,44,45,46],include_baselin:6,inclus:6,incom:2,incompat:46,incomplet:[45,46],incorpor:[21,39],incorrect:[21,29,30,37,46],incorrectli:[21,37],increas:[6,13,37,40],increment:37,indefinit:[11,28],independ:[2,10,46],index:[2,11,25,45,46],indic:[2,11,40,46],indicipher:46,individu:[11,12,21,37,43,46],ineffici:46,infer:37,info:[6,12,37,43],inform:[2,6,11,12,37,44],infrastructur:[37,43,46],ing:2,inherit:[10,12,25,37,43,45,46],init:[2,11,12,15,40,43],init_audio:3,init_cam:[11,13],init_hardwar:[31,37,43],init_log:[2,3,10,37],init_manag:[21,43],init_network:[2,10,11,37],init_pigpio:[3,12],init_sound:[21,25],init_sounds_group:21,init_sounds_individu:21,initi:[2,3,10,11,12,15,21,23,28,37,38,40,44,46],initialis:[17,19],inner_kei:[2,37],input:[2,10,11,12,14,30,40,41,43],inspect:[38,46],instal:[11,13,37,41,44,45,46],install_opencv:11,install_pyspin:11,instanc:[2,11,38,40,43],instanti:[2,3,10,11,12,21,25,31,38,43,46],instead:[2,3,11,12,17,43,46],instruct:[39,46],int32col:[6,28,30,31,43],int_to_float:25,integ:[6,11,25,37,43],integr:[37,45],integrate_dur:[14,43],integrate_fram:13,intend:[2,37,45],intens:45,inter:3,interact:[11,40,46],interest:[37,38],interfac:[23,37,38,40,46],interfer:13,intermediate_node_1:37,intermediate_node_2:37,intern:[2,10,12,21,40,43],internet:44,interpol:[13,37],interpolate_fram:13,interrupt:37,interv:[3,43],intim:46,intra:46,introduc:[42,44],intuit:[38,46],invalid:11,invers:[10,12],inverse_pull_map:12,inverse_trigger_map:12,invert:[12,38],invoc:8,invok:[8,39],ioloop:[2,43],ipv4:[2,3],irrelev:46,is_al:8,is_set:43,is_trigg:[10,12],isdescript:[6,28,30,31,43],isdescriptor:6,isn:11,isoformat:[2,6,11,12,13,28,30,37,43],issu:[24,35,40,45,46],item:[38,43],items:[6,28,30],iter:[21,28,31,38,43],itertool:[2,12,23,28,29,30,43,46],iti:3,its:[2,3,6,10,11,12,17,19,21,25,37,38,39,40,43,44,46],itself:[6,11,12,23],jack2:40,jack:[3,17,23,25,37,40],jack_client:23,jack_last_frame_tim:46,jack_sound:[25,37],jackaudio:[40,46],jackclient:[3,20,22,37,40,46],jackd:[3,17,23,37,40],jackdstr:[3,17],jitter:46,joblib:46,joi:46,join:[2,8,19,41],jonni:[15,25],journalctl:44,json:[2,3,6,15,19,38,39,40,41,43,44,46],json_trick:40,juic:40,just:[2,11,13,21,23,28,36,37,38,40,43,44],keep:[2,3,6,13,21,37,40,43],kei:[2,3,6,8,10,12,25,40,43,44,46],kept:[2,12,23],keyboard:40,keyword:[8,17,19],kick:[40,46],kill:[2,38],know:[2,6,37,43,44],knowledg:11,known:[2,10,38,46],kwarg:[2,3,8,10,11,12,13,21,25,27,28,29,30,31,38,43,46],l_bandwidth:3,l_cal_port:3,l_cal_result:3,l_chang:2,l_child:[2,43],l_clear:[14,43],l_coher:2,l_confirm:2,l_continu:[2,37],l_data:2,l_file:2,l_forward:2,l_handshak:2,l_init:2,l_kill:2,l_measur:[14,43],l_noop:2,l_param:3,l_ping:2,l_process:27,l_start:[2,3,11],l_state:2,l_stop:[2,3,11,14,43],l_stopal:2,l_stream:[2,37],lab:35,label:46,lack:46,lambda:43,larg:[2,12,37],laser:43,last:[6,11,12,21,27,30],last_init_lock:11,last_opencv_init:11,last_stim:21,last_was_correct:21,latenc:[40,46],later:[6,11],latest:46,launch:[12,38,39,40,41,44],launch_pilot:[12,40],launch_termin:41,layer:[38,43,46],ldconfig:40,lead:11,learn:44,learnopencv:40,least:[2,11,27,37,46],leav:[38,40,43,44],led:[3,10,12,14,17,28,30,31,37,38,40,43,44,46],led_rgb:[3,12,17,28,30,40,43],left:[8,21,40,44],leftmost:44,len:[12,38],length:[38,43],less:[21,40,43,46],let:[37,38,40,43,44,46],letter:31,level:[12,19,23,25,31,34,37,41,46],levelnam:[2,3],lib:40,libasound:40,libdir:40,libffi:40,libhdf5:40,liblo:40,libportmidi:40,librari:[9,13,35,40,41,45,46],libreadlin:40,libsamplerate0:40,libsndfile1:40,libx264:11,libzmq:[40,41],lick:43,life:40,light:[28,30,31,38,43],lightli:9,like:[2,6,11,12,21,27,30,31,37,38,39,40,42,43,45,46],limit:[11,38,40,42,44,46],line:[11,12,17,19,40,43],lineag:[40,43],linear:21,linearli:21,link:[15,26,40,46],linux:[39,40],list:[2,3,6,8,11,12,21,27,29,30,31,33,34,37,38,40,43,44,46],list_opt:11,list_spinnaker_camera:11,list_subject:8,listen:[2,3,10,19,37,41,43],listen_port:2,lite:40,liter:46,literatur:44,littl:[23,43],live:2,load:[6,8,23,25,34,37,40,46],load_pilotdb:8,load_subject_data:34,load_subject_dir:34,local:[2,3,11,19,37,40,41],localhost:[2,11],locat:[3,6,19,41],lock:[6,11,23,43],log:[2,3,10,11,15,19,31,35,37,40,41,44,46],log_formatt:[2,3],log_handl:[2,3],log_level:[19,41],logdir:[2,3,10,15,19,41],logfil:46,logger:[2,3,10,31,46],logic:[10,11,12,21,37,38,40,42,43,45,46],login:40,logist:33,loglevel:[15,19,41],look:[3,40],lookup:3,loop:[2,11,23,37],loop_thread:2,lossless:11,lost:46,lot:[37,46],love:[37,46],low:[12,31,37,38,40,44],lower:[40,46],lsm9ds1:37,luigi:46,machin:[11,40,42,46],maco:[39,40,41],made:[2,6,8,11,37,38,46],mag_gain:13,maggain_12gauss:13,maggain_16gauss:13,maggain_4gauss:13,maggain_8gauss:13,magic:37,magnet:13,magnetomet:[13,37],magnific:45,mai:[8,13,37,42,44,46],main:[2,3,31,43,46],maintain:[12,38,43],mainten:41,major:[35,39,46],make:[2,3,6,8,10,11,14,17,19,21,37,38,39,40,42,43,45,46],make_dir:[17,19],make_punish:21,manag:[6,8,10,12,20,23,30,38,40,42,43,45,46],mani:[2,3,8,11,12,37,40,45,46],manpag:[3,17,40],manual:[6,39,41,46],manual_trigg:12,map:[3,8,10,11,12,17,21,28,29,31,40,43,46],march:[35,46],margin:37,mark:[30,31],mask:[3,17],mass:6,match:[2,43],materi:46,matur:46,max:[2,11],maximum:[12,40],mayb:44,mean:[8,11,13,21,37,43,46],meant:37,meantim:46,measur:[3,14,37,43],measure_evt:43,memlock:40,memori:[6,40,46],menu:[37,40,44],mere:44,merg:6,mesa:40,mesh:46,meshnet:46,mess:40,messag:[2,3,10,11,37,40,41,43,44,46],message_cont:37,messi:46,meta:45,metacalss:40,metaclass:[10,11,12,31,37,38,46],metadata:[42,46],meter:37,meth:12,method:[2,3,6,8,10,11,12,13,14,17,19,21,23,25,27,28,29,30,31,37,38,46],mice:[37,38,43],microsd:40,microsecond:[11,12,37,43,45],mid:3,might:[2,25,37,46],mild:30,millisecond:[11,12,43],mimc:46,min_siz:[2,11],mind:40,minim:[2,10,37],minimum:[6,12],minor:[35,46],minu:6,mirror:15,misalign:43,mlx90640:[9,13,37,45],mmap:40,mode:[6,11,12,14,21,30,38,40,43,46],model:[6,14,46],modif:[11,37,45],modifi:[6,9,13,21,37,40,46],modul:[3,12,15,17,19,21,23,24,34,37,38,40,45,46],moment:46,monitor:43,monocl:46,more:[10,11,12,21,37,38,40,42,43,45,46],most:[6,37,38,42,43],mostli:37,motion:[37,45],motiv:45,mous:[14,37,38,43,44],mouse_idx:[14,43],move:[14,29,42,43,44,45,46],move_dtyp:14,movement:[14,38,43],moviepi:40,mp4:11,msg:2,msg_counter:2,msgpack:40,msgport:[3,11,15,17,19,41,43],much:[37,38,43,45,46],muddi:10,multi:11,multihop:46,multipl:[2,3,11,12,23,37,38,42,43,46],multipli:[13,30],multiprocess:[2,11,23,46],must:[2,6,8,10,11,12,13,14,23,25,38,41,43],mutual:12,mysteri:37,n_click:3,n_rep:38,n_trial:[29,44],nafc:[26,44],nafc_gap:[26,30,37],name:[2,3,6,8,10,11,12,15,17,21,23,26,27,31,33,40,43,44,46],name_of_camera_in_task:27,nano:40,nativ:46,natur:44,nchannel:[3,17,23],nctl:40,ncurs:40,ndarrai:[2,11,13,23,25],ndtoverlai:40,ndtparam:40,neater:46,necessari:[10,31,37,40,43,46],necessarili:2,need:[2,3,6,10,11,12,14,17,21,25,29,31,37,38,39,40,43,44,46],needlessli:46,neither:40,nest:[17,19],net:43,net_nod:[2,3,10,11,37,43,45],network:[0,3,10,11,17,37,40,42,43,45,46],neurodata:46,neurosci:[37,45],neutral:[10,14],never:37,neverend:11,new_subject_fil:6,new_valu:6,next:[3,11,21,31,40,43],next_bia:21,next_stim:[21,43],nice:46,nmap:11,node:[2,3,6,10,11,37,40,43,46],nodemap:11,nois:[23,25,30,37,43],noise_amplitud:30,noisier:37,non:[37,38,43],none:[2,3,6,8,10,11,12,13,14,15,17,19,21,23,27,28,30,31,34,38,43],nonzero:30,noop:[27,43],normal:[8,12,37,40,43,44],nosepok:44,note:[2,11,12,39,40,43],noth:[8,25,43],notic:44,now:[13,21,23,30,31,37,38,41,42,43,45,46],nowtim:43,nperiod:40,npsappmanag:[17,19],npyscreen:[17,19,40,41,46],ntp:46,ntrial:29,nudg:37,num_stag:28,number:[2,3,6,8,10,11,12,13,17,21,23,28,29,30,31,37,38,40,42,43,45,46],numer:[8,11,43],numpi:[2,11,13,15,23,25,38,40,41,43],numpy_arrai:2,nut:10,object:[2,3,6,8,10,11,12,13,14,17,21,25,27,28,29,31,37,38,40,42,43,44,46],obnoxi:11,observ:46,obviou:[37,46],obvious:40,occur:[8,31],octob:35,odict:[31,43],off:[10,12,14,28,31,38,40,43,46],offroad:2,offset:37,offtim:38,often:[2,46],old:[6,37],oldest:46,omit:38,on_recv:2,onboard:40,onc:[2,11,12,31,37,40,43,44,46],one:[2,6,12,13,21,23,28,37,38,40,42,43,44,45,46],ongo:[42,43],onli:[2,6,10,11,12,23,27,31,37,39,40,43,44,46],onstart:[17,19],opac:24,opaqu:46,open:[2,3,6,11,12,28,30,31,37,40,43],open_dur:3,open_fil:3,open_hdf:6,opencl:46,opencv:[11,37,40,45,46],opencv_last_init_tim:11,openephi:46,opengl:40,oper:[8,10,11,27,37,39,40,41,42,43,44,45,46],opportun:40,oppos:38,optim:46,option:[2,3,8,11,37,38,39,41,43,44,46],order:[2,6,10,11,38,43,44,46],ordereddict:31,org:[36,42,46],organ:21,orient:37,origin:2,other:[2,6,10,11,12,13,15,21,25,27,37,38,40,43,46],otherwis:[2,3,6,11,12,21,23,31,43,46],our:[2,3,6,13,17,19,21,38,40,41,43,44,45,46],ourselv:3,out:[6,8,12,21,37,40,44,45,46],outbox:2,outcom:21,outport:23,output:[3,6,10,11,12,13,17,23,27,40,43,45],output_filenam:11,outright:46,outsid:[3,30,43],over:[2,11,12,13,17,19,21,28,29,31,37,40,42,43,46],overhaul:46,overlai:40,overlap:[11,37,38],overrid:[2,8,12,17,19,37],overridden:11,overrun:40,overview:45,overwrit:[11,38],overwritten:[11,12],own:[2,3,10,11,12,17,21,37,39,40,42,43,44,45,46],p16:[3,17,40],p75:[3,17,40],pack:2,pack_arrai:[2,11],packag:[23,38,40,41,46],pad:25,page:[37,42,45],pair:[2,37,43],panda:[6,8,33,40,41],panel:44,parallax:26,param1:27,param:[2,6,8,12,15,17,19,25,27,28,29,30,31,33,41,44,46],param_1:44,paramet:[2,3,6,8,10,11,12,13,14,15,17,19,21,23,24,25,27,28,29,30,31,33,34,37,38,40,42,43,44,46],parameter:[12,38,40,46],parent:[25,40,43,46],parent_pilot:43,parentapp:17,parentid:43,parentip:43,parentport:43,parlanc:12,parsimoni:46,part:[39,45],partial:6,particular:[6,10,11,21,44,46],particularli:[37,42,46],pass:[2,3,6,8,10,11,12,13,23,25,30,43,46],passag:43,passwd:40,password:[40,44],past:[6,29],past_protocol:6,patch:46,path:[2,3,6,11,15,17,19,25,34,40,41],pattern:46,pcm512x:40,pcm:40,per:[2,23,43,46],percent:[21,29],percentag:43,perfect:37,perform:[3,11,17,19,37,38,39,42,43,45,46],period:[2,40],permiss:[17,19],persist:42,phase:6,physic:10,pi3:40,pick:37,pickl:46,piec:[2,38,46],piecem:40,pig:[12,38],pigpio:[3,9,10,12,13,17,31,37,40,43,45,46],pigpiod:[10,12,38,46],pigpiomask:[3,17],pigs_funct:12,pillow:40,pilot:[0,2,10,17,19,28,37,38,39,41,42,43,46],pilot_button:46,pilot_db:[8,15,19,41,44,46],pilot_st:[2,3,43],pilotsetupform:[17,46],pimoroni:37,pin2:31,pin:[3,10,12,13,17,28,30,31,37,38,40,43],pin_bcm:12,pin_id:31,ping:2,pinout:10,pip3:41,pip:[40,46],pipe:23,pixel:37,place:[2,12,43],plai:[23,25,30,37,40,43],plaintext:37,plan:[37,45,46],play_continu:37,play_evt:23,play_punish:[21,43],playback:[40,43],pleas:[11,45],plot:[0,2,28,30,31,33,37,42,44,46],plot_list:43,plot_psychometr:33,plot_tim:2,plots_menu:37,plug:31,plugin:46,png:11,pngoption:11,point:[8,28,30,31,37,43,46],pointer:11,poissonian:3,poke:[3,17,28,30,31,38,40,43,44],polar:[12,37,38,40],poll:46,pool:[40,46],pop:2,pop_dialog:37,popul:[11,19,29,41,46],port:[2,3,11,12,17,19,21,23,28,30,31,40,41,43,44],port_calibr:12,port_id:31,port_nam:3,portal:46,pos:[6,28,30],posit:[10,37],possibl:[11,12,37,43,46],post:[6,46],postgresql:46,potenti:[37,46],practic:[37,43],pre:6,preambl:38,prebuilt:11,preciou:38,precis:[12,25,37,43,45,46],precomput:43,predict:46,pref:[2,3,6,10,11,12,19,23,25,31,37,38,39,40,41,44,45,46],prefdict:15,prefer:[12,15,37,40],prefix:43,prefs_fil:[3,15],prei:44,preinstal:39,prepar:[6,11,39,43],prepare_messag:2,prepare_run:6,prerequisit:39,present:[3,8,21,37,40,42,43,44,46],presetup:[39,41],presetup_pilot:[40,46],presetup_termin:41,press:[44,46],pretti:43,prettyprint:46,prevent:[37,38],previou:[6,12,37,38,43],previous:[37,38,40,46],primarili:[10,46],primit:43,principl:38,print:[11,44],prior:21,prioriti:[3,40],privileg:40,problem:[37,46],process:[2,3,11,23,37,38,39,43],produc:[37,40,43,46],product:[11,13],product_id:14,program:[44,46],programm:[17,19],programmat:43,progress:[3,11,45],project:46,prompt:[17,19,40],proper:37,properli:6,properti:[11,12,13,37,46],proport:[11,12,21,25,30,43],protocol:[6,12,15,19,26,29,37,41,42,46],protocol_nam:6,protocol_script:44,protocoldir:[6,15,19,41],provid:[2,11,12,37,41,42,43],psychometr:[32,37],psychopi:40,psychopy3:40,pub:46,publish:46,pull:[3,6,11,12,17,23,38,40],pull_map:12,pull_ud:38,pulldown:[3,12,17],pullup:[3,12,17,37],puls:[11,12,38,43],pulse_width:12,punish:[30,31,43],punish_block:[31,43],punish_dir:[30,31],punish_dur:[30,43],punish_stim:[30,31,43,44],purpos:[10,43,46],push:[2,40,46],push_id:2,push_ip:2,push_outbox:2,push_port:2,pusher:2,pushport:[3,15,17],put:[6,11,13,29,43],pwm:[12,40],pyglet:40,pyo:[3,17,24,25,40],pyo_serv:[3,24],pyo_sound:25,pyopengl:40,pyoserv:[3,20,22],pyparallel:40,pyper:46,pyqtgraph:[37,41],pyseri:40,pysid:[37,41,46],pyside2:[37,41,42],pyspin:[11,37],pytabl:[6,43],python2:40,python3:[40,41],python:[2,3,10,12,13,37,38,40,41,45,46],pyyaml:40,pyzmq:[40,41],q_lock:23,qt4:41,qt5:[37,41,42,46],qtcore:37,qtimer:37,queri:11,quest:37,queue:[2,6,11,23,37,43,46],queue_siz:11,quickli:37,quit:[23,40,46],quit_evt:[23,43],radio:46,raft:37,rais:[8,11,40],random:[21,43],randomli:[21,28,30],randomnumbersandlett:40,rang:[6,12,13,43],rapid:38,rare:21,rasbian:39,raspberri:[3,10,11,37,38,40,42,43,44,45],raspberrypi:40,raspbian:39,raspi:[12,37,40],rat:[12,37,40],rate:[2,3,17,21,23,27,37,40,43],rather:[2,9,11,12,23,25,37,38,40,43,46],rdisk:40,reach:30,read:[6,11,13,37,38,43,46],readabl:[11,43,46],readable_attribut:11,readi:43,readthedoc:46,real:12,realli:[2,40],realtim:[40,46],reason:40,reassign:[2,6],reboot:40,rebuilt:[37,45],receipt:37,receiv:[2,6,37,40,43,44],recent:[6,43,46],recipi:[11,37,46],recommend:[6,11,40,43],reconfigur:[40,46],reconstitut:2,reconstruct:6,record:[11,12,25,37],record_ev:12,recov:46,recurs:[17,19,38,46],recv:2,red:[12,31,43],redefin:[10,14,21,38],reduc:[37,40],redund:43,refactor:[6,37,38,45],refer:[5,11,12,29,30,31,43],reflect:[21,37,46],refrain:43,refresh:44,regimen:44,region:46,regist:[23,38,43,44],register_task:46,registri:46,regress:33,reinforc:30,rel:[2,6,11,25,27,37,44,46],rel_i:38,rel_x:38,relat:44,releas:[2,3,10,11,12,13,14,21,27,28,31,36,37,38,43,46],relev:[2,11,38,44,46],reli:[12,37,43,46],reliabl:[40,46],remain:[11,12,23,27,38,44,46],remov:37,renam:37,reorgan:37,repeat:[2,12,28,37],repeat_interv:2,repeatedli:[3,43],repetit:12,replac:[37,40,46],replic:[45,46],repo:[19,37,41],repo_dir:15,repo_directori:40,repodir:[15,19,41],report:[27,28,37,43,44,46],repositori:[6,9,15],repres:[8,12,46],represent:6,req_reward:[30,31,43,44],request:[2,3,6,11,30,31,37,44],request_reward:44,requir:[2,3,10,11,17,37,38,39,40,41,43,44,46],requirements_doc:37,requirements_termin:37,resampl:25,rescal:25,resembl:46,resend:[2,37],resent:2,reserv:8,reshap:40,resistor:[12,37,38],resiz:[40,46],resolut:[13,37],resourc:[10,11,14,37,38,43,46],respect:[8,11,12,37],respond:[2,21,30,43,44],respons:[21,28,30,31,33,40,43],rest:[2,3],restart:46,restructur:37,result:[3,12,27,37,43,46],resum:37,retriev:[11,12,44,46],return_id:27,returnthread:8,revers:[8,31,40,43],revis:11,reward:[12,28,30,31,37,40,43,44,45],reward_manag:44,rfs:[3,17,40],rgb:[12,38,40],rhythm:46,ride:[17,19],right:[8,21,23,40,44],rightmost:44,ring:13,rise:12,rising_edg:12,risk:[6,46],roadmap:45,robust:46,role:39,roll:[31,43,46],roll_span:34,roll_typ:34,roll_window:[30,31,43],rollmean:[30,31,43],root:[6,17,19],roughli:46,round:[38,43],rout:[2,46],route_port:2,router:[2,3,17,19,41,46],routin:[3,11,15,37,38,39,40,41,42,43,46],row:[3,29,43],rpi3:46,rpi:[12,40],rq_timestamp:[30,31,43],rsa:[40,44],rtprio:40,rtype:[2,11],rule:21,run:[2,3,6,8,10,11,12,17,19,23,31,37,38,40,41,42,43,46],run_task:3,runtim:[39,46],runtimeerror:8,rw_mode:11,ryanjdillon:15,s26:28,sadon:2,safe:[10,14],same:[2,6,8,11,12,13,21,37,38,40,43],sampl:[3,17,23,37,40,46],sarg:46,sargehttp:46,save:[2,3,6,11,37,38,43,44,46],save_data:6,scale:[13,14,18,30,45],scaling_governor:40,schema:[21,44],scheme:[10,46],scikit:41,scipi:[13,25,41],scl:13,scrappi:21,scratch:[43,46],screen:[40,46],screenshot:43,screwi:23,script:[10,11,12,13,17,38,39,45,46],script_:12,script_count:12,script_id:12,scroll:46,sda:13,sdk:[11,37,45],seamlessli:37,search:45,second:[8,11,13,37,43,44],second_group:21,section:[38,43,44,46],secur:40,sed:40,see:[3,5,6,10,11,12,13,17,19,30,35,37,38,39,40,41,42,43,44,45,46],seem:[11,46],segment:[30,31,43],select:[2,3,6,11,21,33,44,46],self:[6,12,21,30,31,38,43],semant:46,send:[2,3,11,12,37,43,44,46],send_outbox:2,sender:[2,27,37,46],sens:46,sensibl:45,sensor:[13,37,43,44,45],sent:[2,37,43],sent_plot:2,separ:[11,13,43,46],sequenc:38,sequenti:8,seri:[11,12,38,43],serial:[2,6,11,37,44],serializ:2,serialize_arrai:2,series_:12,series_thread:38,serv:[13,27,43],server:[3,17,23,24,25,40,46],server_typ:25,servic:[10,17,39,40,41,46],session:[6,37],set:[2,3,6,9,10,11,12,13,17,19,21,23,30,31,37,38,39,40,41,43,44,46],set_color:[38,43],set_l:[31,43],set_mod:38,set_pull_up_down:38,set_pwm_dutycycl:38,set_reward:[31,43],set_trigg:[21,43],set_weight:6,setattr:43,setter:[37,46],setup:[3,10,11,13,15,18,37,38,39,42,43,44,45,46],setup_mlx90640:13,setup_pilot:[3,10,12,16,40,46],setup_pilot_:44,setup_scal:16,setup_termin:[16,41,42],setupapp:[17,19],setupform:[17,19],setuptool:46,setvalu:11,setverbos:24,sever:[37,43],shade:43,shape:[6,11,13,28,30,37,42,44],shape_sensor:13,share:[42,46],shared_memori:46,shift:43,shitload:2,shock:46,shorter:46,shorthand:12,should:[2,3,6,8,10,11,12,15,17,19,21,23,25,27,29,30,37,38,39,40,43,44,46],shouldn:[12,43,46],shovel:37,show:12,shown:43,shut:[28,43],side:[21,44],signal:[2,3,6,11,14,23,28,30,31,37,38,43,46],silent:[23,25,37],similar:[6,37,38,43,46],similarli:[21,38],simpl:[6,11,37],simpli:43,simplifi:[2,37,38,41],simultan:6,sin:25,sinc:[11,31,37,38,40,43,46],sine:25,singl:[2,6,11,12,23,37,38,43,45],single_dur:38,sink:46,size:[37,43,46],skew:21,skvideo:11,sleep:[3,38],sleev:38,slight:37,slightli:40,slot:[37,46],slow:43,slower:40,small:8,smaller:46,smart:38,smarter:[2,46],smooth:39,smoothli:43,snd_rpi_hifiberry_dacplu:40,sndrpihifiberri:[3,17,40],sneaker:[12,37,40],sock:2,socket:[2,11,12,37],soft:40,softwar:[10,46],solenoid:[3,12,17,28,30,31,40,43,44],solidifi:43,solut:10,solv:46,some:[2,10,11,12,15,21,23,29,30,31,37,38,40,43,45,46],some_index_for_your_mous:38,someth:[40,46],sometim:37,somewhat:46,song:2,sorri:[38,40],soul:40,sound:[3,15,19,20,21,23,30,37,40,41,42,43,44,46],sound_dict:21,sound_info:43,sound_list:[21,25],sound_param:21,sound_stim:21,sound_widget:43,soundcard:40,sounddevic:40,sounddir:[2,15,19,25,41],soundfil:40,sourc:[2,3,6,8,10,11,12,13,14,15,17,19,21,23,24,25,27,28,29,30,31,33,34,46],sparkfun:13,spawn:[2,6,10,11,13,38,46],speak:40,speaker:[25,40,46],specif:[12,37,38,43,44,46],specifi:[2,3,8,12,28,31,37,38,40,43,44,46],speech:[25,44],speed:[37,40],spill:12,spinnak:[11,37,45],spiral:46,splash:3,split:[2,37,40,46],splitform:[17,19],sprawl:46,ssh:[40,44],stack:44,stackoverflow:8,stage:[3,6,12,28,29,30,31,42,46],stage_block:[3,27,28,30,31,43],stage_list:43,stage_nam:[27,28,30,31,43],stai:6,stamp:14,standalon:3,standard:[8,37,46],start:[2,3,6,8,10,11,14,15,17,23,24,27,28,29,30,36,37,38,40,43,44,46],start_now:27,start_plot_tim:2,startup:[40,46],stash:[2,6,11,43],stash_curr:6,state:[2,3,6,10,11,12,14,37,38,42,44,46],station:[2,3,40,43],stationari:23,statu:[2,6,12,21,37,40,46],stdout:[11,37],steadi:[14,43],step:[6,33,34,37,39,44,46],step_data:34,step_n:6,step_nam:[33,44],step_view:34,still:[8,11,38,40,42,43,46],stim:[21,25,30,42,43,44],stim_end:[30,43],stim_light:30,stim_manag:[21,37,43],stim_start:[30,43],stimuli:[21,30,37,40,43,44,45,46],stimulu:[21,30,31,37,42,43,44,45,46],stimulus_manag:44,stolen:[15,40],stop:[2,3,6,10,11,12,13,14,23,27,30,38,43,46],stop_evt:23,stop_run:6,stop_script:12,stopal:2,storag:[2,6,42,43,46],store:[3,6,12,13,15,19,21,23,31,38,40,41,43,44,46],store_group:21,store_script:12,store_seri:12,stored_color:38,str:[2,3,6,8,10,11,12,13,15,17,19,23,25,27,33,34,38],straightforward:38,stream:[2,10,11,27,37,40,45,46],streamlin:46,streamz:46,strict:38,strictli:[40,46],string:[3,6,8,11,12,17,25,26,29,37,38,40,43,44],string_of_camera_class:27,string_param:25,stringcol:[6,28,30,31,43],strongli:[40,44],structur:[6,14,21,35,38,40,41,43,44,45,46],stuck:[44,46],style:[0,34,37,46],stylesheet:[5,37],sub:[2,27,46],subclass:[2,8,11,21,25,40,43,46],subconfigur:43,subdevic:40,subdict:31,subdictionari:43,subject:[0,2,3,8,21,28,30,33,34,37,42,43,45,46],subject_id:33,subject_list:46,subject_nam:34,subject_posit:46,subject_protocol:33,submit:[23,45],submodul:[9,37],subset:34,substanti:37,success:12,sudo:40,suffici:10,suit:46,suitabl:11,sum:21,summar:34,summat:11,sunk:40,superclass:[11,12,13],superfast:37,support:[11,23,37,43,44,45],suppos:44,sure:[2,6,8,11,38,40],surprisingli:37,swarm:[40,42],synchron:[11,12,37,46],synthesi:40,sys:[11,40],system:[2,8,9,10,11,12,14,37,38,39,40,41,42,43,44,46],systemctl:40,systemd:[10,17,38,40,41],systemwid:[13,19,41],t2000:[3,17,40],tabl:[2,3,6,25,28,29,30,31,37,40,41,43,44,46],tableextens:[3,29],tag:[8,27,28,30,31,43,44],take:[2,10,11,12,38,40,43],taken:[8,10,38,43],target:[8,21,28,30,31,38,43],task:[2,3,6,8,10,12,14,17,27,28,29,30,37,38,40,41,45,46],task_class:3,task_list:[3,26,44,46],task_nam:6,task_param:3,task_typ:[3,43,44],taught:44,teach:44,tee:40,tell:[2,31,38,43],temperatur:[13,37,45],templat:42,tempor:44,temporari:37,temporarili:46,tensor2tensor:46,term:[44,46],termin:[0,2,3,8,11,17,19,23,28,33,37,39,40,42,43,44,46],terminal_ip:44,terminal_network:37,terminal_st:2,terminalip:[3,11,15,17],terminalport:11,terminalsetupform:19,terminolog:37,test:[3,11,26,37,39,40,43,45,46],test_dlc_lat:26,than:[2,9,10,11,12,21,23,25,37,38,40,43,46],thank:45,thei:[2,11,12,21,23,25,31,37,38,39,43,44,46],them:[2,3,6,10,12,21,37,38,40,42,43,44,46],thereof:8,thi:[2,3,6,8,10,11,12,13,14,17,19,21,23,25,29,31,34,37,38,40,42,43,44,45,46],thin:43,thing:[21,37,39,43,46],thingivers:46,think:37,thortex:46,those:[38,40,46],though:[2,43],thread:[2,3,6,8,10,11,12,13,28,30,31,37,38,43,46],threaded_color_seri:38,threaded_loop:2,threadgroup:8,threadsaf:10,three:[11,12,37,38,43,44],thresh:[14,21,27,43],thresh_trig:[14,43],thresh_typ:[14,43],thresh_val:14,threshold:[14,21,27,29,30,43,44],thresholded_linear:[21,30],through:[2,6,8,11,12,28,37,38,40,43,44,46],throughout:[37,46],throughput:37,thu:[2,11,37,39,40,42,43,46],thx:8,tick:[9,12,31,37,43,46],ticks_to_timestamp:37,time:[2,3,6,8,11,12,13,14,21,23,29,37,38,40,41,43,44,45,46],timeout:[8,12,40,43],timer:[2,38,43],timestamp:[2,3,6,9,10,11,12,13,14,28,37,38,43,45,46],to_csv:[6,37],toggl:12,token:25,told:[2,43],tone1:21,tone1_param:21,tone2:21,tone2_param:21,tone3:21,tone3_param:21,tone4:21,tone4_param:21,tone:[21,25,30,44],too:[10,40,46],tool:[34,40,45,46],tooltip:46,top:46,topolog:2,tornado:[2,40,41,43],torrent:40,tostr:11,total:[12,14],toward:37,tqdm:[11,41],trace:44,track:[43,46],train:[6,38,43,45],transfer:[3,40,46],transform:[8,12,26,27,46],transit:[12,37,38,42],translat:[37,46],transmiss:45,transmit:43,transpar:[38,43],treat:46,tree:[2,46],treelik:40,tri:[8,12],trial:[3,6,21,28,29,30,31,34,37,42,43,44,46],trial_count:[28,30,43],trial_data:[6,31],trial_end:[6,30,43],trial_manag:46,trial_num:[6,28,30,31,43],trial_view:32,trialdata:[6,28,30,31,43,46],trialwis:37,trick:38,trig_fn:21,trigger:[10,11,12,14,23,27,28,29,30,31,37,38,40,43,46],trigger_edg:12,trigger_fn:[10,14],trigger_map:[12,38,46],trigger_ud:38,trivial:[3,40,43],troubleshoot:45,ttl:[2,11,12],tup:33,tupl:[6,8,11,12,13,21,33,38],turn:[10,12,14,28,30,31,38,40,42,43,46],tweak:46,twice:13,two:[2,6,9,10,11,12,28,30,37,39,40,43,44],txt:[13,37,40],type:[2,3,6,8,10,11,12,13,14,15,17,21,23,25,27,28,29,30,31,33,37,38,40,43,44,46],typic:[2,3,6,10,12,14,19,33,37,38,41],ubuntu:39,uli:42,ultim:46,umount:40,unambigu:[37,46],unavail:46,unchang:37,unclear:46,uncompress:11,undecid:46,under:[40,42,43],underrun:40,undocu:[27,34],unfinish:34,unfold:[17,19],unfold_valu:[17,19],unfortun:43,unhandl:8,unicod:15,unif:44,unifi:[2,21,39,43,46],unimport:46,uniqu:[2,8,40,43],unit:[12,44,46],univers:[40,44],unix:39,unless:[2,6,31,46],unlik:[2,12],unload:[10,14],unmount:40,unmountdisk:40,unnecessari:25,unpack:[38,43],unpickl:46,unproduct:46,until:[2,8,11,21,23,27,37,40,43],unus:[11,23,25],unzip:40,upcom:27,updat:[2,6,12,14,19,21,29,30,35,38,40,41,43,44],update_biographi:6,update_dur:43,update_histori:6,update_protocol:37,update_st:3,update_weight:6,upgrad:[40,41,46],upon:[2,12,23],upstream:[2,3,17,40,43],upstream_ip:2,url:37,usb:[10,27,37,40,43],use:[2,3,6,10,11,12,13,17,19,33,37,38,40,41,42,43,45,46],use_histori:6,used:[2,3,6,10,11,12,14,17,19,23,25,26,28,30,37,38,40,41,42,43,45,46],user:[11,12,15,17,19,37,40,41,42,43,44,45],usernam:40,userpref:40,uses:[2,9,10,11,12,13,14,21,37,38,40,41,43,46],using:[2,3,12,13,14,21,25,30,37,40,43,44,46],usr:[3,15,17,19,40,41],usual:[6,10,33,43],util:[0,40],v4l2:11,v4l_info:11,val:11,valid:2,valu:[2,3,6,8,11,12,13,14,15,17,19,21,27,30,31,38,43,44,46],value_1:44,valv:[3,12],var_i:33,var_x:33,vari:12,variabl:[2,3,6,10,11,12,13,15,21,23,25,28,29,30,31,33,37,38,43,46],variat:30,vc4cl:46,vcc:13,vel:14,vel_tot:[14,43],veloc:[27,43],vendor_id:14,venv37:15,venv:15,verbos:8,veri:[10,37,38,40,45,46],versa:[12,25],version:[2,6,9,11,12,13,36,37,38,39,40,41,43,44,46],vertic:[11,37],via:[2,21,40,43,44],vice:[12,25],video4linux:11,video:[2,11,19,26,37,39,41,45,46],video_child:[26,27,37],video_writ:11,videocaptur:11,videox:11,vidioc_qbuf:11,view:[6,46],viola:43,virtu:44,virtual:46,vision:[11,37],visual:[19,37,40,41,42,43,45,46],visul:34,viz:[15,37],vizdir:[15,19,41],vol:[12,31],voltag:[12,43],volum:[3,12,31,44,45],vowel:25,waf:40,wai:[12,37,40,46],wait:[3,8,11,13,23,28,37,38,43],wall:44,want:[2,40,43,46],warn:[10,14,37],wasn:21,watchdog:12,water:[3,12,26,28,30,40,44],wav:[2,25],wave:25,wavfil:25,web:46,webcam:[11,37],wehr:35,weigh:14,weight:[6,21],weight_tabl:6,well:[12,37,41,43,44,46],were:[37,38,43],what:[2,3,6,10,11,12,21,30,31,36,37,38,43,46],whatev:40,wheel:[14,26,27,38,40],wheel_0:43,wheel_:43,wheel_child:[26,27,43],when:[2,3,6,8,10,11,12,13,14,21,23,27,28,30,31,37,38,40,43,44,46],whenev:[6,12,37,43,46],where:[2,6,8,21,27,37,43,44,46],whether:[6,8,11,12,14,21,28,30,34,38,40,43,46],which:[2,3,6,10,11,12,13,17,21,25,30,34,37,38,39,40,41,42,43,44,46],whistl:38,white:[25,30,37,43],whitepap:[37,40,42,43,45],who:46,whole:[6,37],wholli:45,whose:[8,12,43],widget:[5,17,19,44],widget_list:17,width:[11,12,38,46],window:[19,21,29,31,37,41,44,46],wintyp:40,wire:2,within:[2,39,40,43,44,46],without:[2,6,10,12,37,38,44,46],wizard:[40,46],won:[10,43],work:[3,10,11,13,14,43,45,46],workaround:11,world:3,worth:46,would:[2,6,8,12,37,46],wouldn:37,wrap:[12,37],wrapper:[11,12,40,43],writable_attribut:11,write:[6,11,12,40,42,45,46],writeabl:11,writeable_attribut:11,writer:[11,23],written:[11,37,38,43,46],wrote:37,wth:11,www:[11,13,36,40,42],x11:40,x264:[11,37],x_vel:43,xinit:40,xml:11,xorg:40,xserver:40,y_vel:43,yes:[38,40,44,46],yet:[11,18,37,43,46],yield:[21,43,46],you:[2,8,11,12,13,21,37,38,40,43,44,45],young:45,your:[11,40,44,45,46],your_disk:40,zero:[23,25,28,29],zero_arr:23,zeromq:[2,40,41],zip:38,zmq:[2,37,41,46],zyre:46},titles:["core","gui","networking","pilot","plots","styles","subject","terminal","utils","external","hardware","cameras","gpio","i2c","usb","prefs","setup","setup_pilot","setup_scale","setup_terminal","stim","managers","sound","jackclient","pyoserver","sounds","tasks","children","free_water","graduation","nafc","task","viz","psychometric","trial_viewer","Changelog","v0.2.0 (October 26, 2019)","v0.3.0 (March 3rd, 2020)","Writing a Hardware Class","Installation","Pilot Setup","Terminal Setup","Overview","Writing a Task","Training a Subject","Autopilot - Distributed Behavior","To-Do"],titleterms:{"3rd":37,"class":[38,43,46],"new":[37,45],The:43,Using:[43,44],addit:43,agent:43,attribut:43,audio:40,autopilot:46,beambreak:38,behavior:46,bonsai:46,bug:46,bugfix:37,camera:11,changelog:35,child:43,children:27,close:46,code:37,com:46,complet:46,connect:44,core:0,creat:44,creation:44,data:43,dblink:46,debug:44,deeplabcut:46,discrim:43,discuss:46,distribut:43,ephi:46,extern:[9,37],featur:37,forum:46,four:43,free_wat:28,googl:46,gpio:[12,38],graduat:29,group:46,gui:1,hardwar:[10,38,43],high:46,href:46,http:46,i2c:13,improv:[40,46],indic:45,initi:43,input:38,instal:[39,40],integr:46,jackclient:23,led_rgb:38,librari:37,loop:46,low:46,lowest:46,major:37,manag:21,manual:[40,44],march:37,medium:46,method:43,minor:37,modul:42,nafc:[30,43],network:2,octob:36,open:46,option:40,output:38,overview:42,param:43,parameter:43,perform:40,pigpio:38,pilot:[3,40,44],pipelin:46,plot:[4,43],pnfrhdtza:46,pref:[15,43],preinstal:40,prerequisit:41,prioriti:46,process:46,program:42,protocol:44,psychometr:33,pyoserv:24,rasbian:40,raspbian:40,reinforc:43,request:43,roadmap:46,run:44,script:[40,41],setup:[16,40,41],setup_pilot:17,setup_scal:18,setup_termin:19,smart:43,sound:[22,25],span:46,stage:43,step:40,stim:20,structur:[37,42],style:5,subject:[6,44],tabl:45,task:[26,31,42,43,44],termin:[7,41],todo:[2,3,6,11,21,25,38,39,40,42,43,44],topic:46,tour:42,train:44,trial_view:34,updat:37,usb:[14,38],user:46,util:8,veri:43,version:35,video:40,vision:46,viz:32,what:45,wheel:43,wizard:44,write:[38,43]}})
+Search.setIndex({docnames:["autopilot.core","autopilot.agents.gui","autopilot.core.networking","autopilot.agents.pilot","autopilot.agents.plots","autopilot.agents.styles","autopilot.core.subject","autopilot.agents.terminal","autopilot.core.utils","autopilot.external","autopilot.hardware","autopilot.hardware.cameras","autopilot.hardware.gpio","autopilot.hardware.i2c","autopilot.hardware.usb","autopilot.prefs","autopilot.setup","autopilot.setup.setup_pilot","autopilot.setup.setup_scale","autopilot.setup.setup_terminal","autopilot.stim","autopilot.stim.managers","autopilot.stim.sound","autopilot.stim.sound.jackclient","autopilot.stim.sound.pyoserver","autopilot.stim.sound.sounds","autopilot.tasks","autopilot.tasks.children","autopilot.tasks.free_water","autopilot.tasks.graduation","autopilot.tasks.nafc","autopilot.tasks.task","autopilot.viz","autopilot.viz.psychometric","autopilot.viz.trial_viewer","changelog/index","changelog/v0.2.0","changelog/v0.3.0","guide.hardware","guide.installation","guide.installation.pilot","guide.installation.terminal","guide.overview","guide.task","guide.training","index","todo"],envversion:{"sphinx.domains.c":2,"sphinx.domains.changeset":1,"sphinx.domains.citation":1,"sphinx.domains.cpp":3,"sphinx.domains.index":1,"sphinx.domains.javascript":2,"sphinx.domains.math":2,"sphinx.domains.python":2,"sphinx.domains.rst":2,"sphinx.domains.std":1,"sphinx.ext.intersphinx":1,"sphinx.ext.todo":2,"sphinx.ext.viewcode":1,sphinx:56},filenames:["autopilot.agents.rst","autopilot.agents.gui.rst","autopilot.agents.networking.rst","autopilot.agents.pilot.rst","autopilot.agents.plots.rst","autopilot.agents.styles.rst","autopilot.agents.subject.rst","autopilot.agents.terminal.rst","autopilot.agents.utils.rst","autopilot.external.rst","autopilot.hardware.rst","autopilot.hardware.cameras.rst","autopilot.hardware.gpio.rst","autopilot.hardware.i2c.rst","autopilot.hardware.usb.rst","autopilot.prefs.rst","autopilot.setup.rst","autopilot.setup.setup_pilot.rst","autopilot.setup.setup_scale.rst","autopilot.setup.setup_terminal.rst","autopilot.stim.rst","autopilot.stim.managers.rst","autopilot.stim.sound.rst","autopilot.stim.sound.jackclient.rst","autopilot.stim.sound.pyoserver.rst","autopilot.stim.sound.sounds.rst","autopilot.tasks.rst","autopilot.tasks.children.rst","autopilot.tasks.free_water.rst","autopilot.tasks.graduation.rst","autopilot.tasks.nafc.rst","autopilot.tasks.task.rst","autopilot.viz.rst","autopilot.viz.psychometric.rst","autopilot.viz.trial_viewer.rst","changelog/index.md","changelog/v0.2.0.rst","changelog/v0.3.0.rst","guide.hardware.rst","guide.installation.rst","guide.installation.pilot.rst","guide.installation.terminal.rst","guide.overview.rst","guide.task.rst","guide.training.rst","index.md","todo.rst"],objects:{"autopilot.core":{networking:[2,0,0,"-"],pilot:[3,0,0,"-"],styles:[5,0,0,"-"],subject:[6,0,0,"-"],utils:[8,0,0,"-"]},"autopilot.core.networking":{Message:[2,1,1,""],Net_Node:[2,1,1,""],Pilot_Station:[2,1,1,""],Station:[2,1,1,""],Terminal_Station:[2,1,1,""],serialize_array:[2,4,1,""]},"autopilot.core.networking.Message":{__contains__:[2,2,1,""],__delitem__:[2,2,1,""],__getitem__:[2,2,1,""],__setitem__:[2,2,1,""],_serialize_numpy:[2,2,1,""],changed:[2,3,1,""],expand:[2,2,1,""],flags:[2,3,1,""],get_timestamp:[2,2,1,""],id:[2,3,1,""],key:[2,3,1,""],sender:[2,3,1,""],serialize:[2,2,1,""],serialized:[2,3,1,""],timestamp:[2,3,1,""],to:[2,3,1,""],ttl:[2,3,1,""],validate:[2,2,1,""],value:[2,3,1,""]},"autopilot.core.networking.Net_Node":{context:[2,3,1,""],get_stream:[2,2,1,""],handle_listen:[2,2,1,""],id:[2,3,1,""],init_logging:[2,2,1,""],init_networking:[2,2,1,""],l_confirm:[2,2,1,""],l_stream:[2,2,1,""],listens:[2,3,1,""],log_formatter:[2,3,1,""],log_handler:[2,3,1,""],logger:[2,3,1,""],loop:[2,3,1,""],loop_thread:[2,3,1,""],outbox:[2,3,1,""],port:[2,3,1,""],prepare_message:[2,2,1,""],release:[2,2,1,""],repeat:[2,2,1,""],repeat_interval:[2,3,1,""],send:[2,2,1,""],sock:[2,3,1,""],threaded_loop:[2,2,1,""],timers:[2,3,1,""],upstream:[2,3,1,""]},"autopilot.core.networking.Pilot_Station":{l_change:[2,2,1,""],l_child:[2,2,1,""],l_cohere:[2,2,1,""],l_continuous:[2,2,1,""],l_file:[2,2,1,""],l_forward:[2,2,1,""],l_noop:[2,2,1,""],l_ping:[2,2,1,""],l_start:[2,2,1,""],l_state:[2,2,1,""],l_stop:[2,2,1,""]},"autopilot.core.networking.Station":{child:[2,3,1,""],ctx:[2,3,1,""],get_ip:[2,2,1,""],handle_listen:[2,2,1,""],id:[2,3,1,""],init_logging:[2,2,1,""],ip:[2,3,1,""],l_confirm:[2,2,1,""],l_stream:[2,2,1,""],listen_port:[2,3,1,""],listener:[2,3,1,""],listens:[2,3,1,""],log_formatter:[2,3,1,""],log_handler:[2,3,1,""],logger:[2,3,1,""],loop:[2,3,1,""],prepare_message:[2,2,1,""],push:[2,2,1,""],push_id:[2,3,1,""],push_ip:[2,3,1,""],push_outbox:[2,3,1,""],push_port:[2,3,1,""],pusher:[2,3,1,""],release:[2,2,1,""],repeat:[2,2,1,""],repeat_interval:[2,3,1,""],routes:[2,3,1,""],run:[2,2,1,""],send:[2,2,1,""],send_outbox:[2,3,1,""],senders:[2,3,1,""],timers:[2,3,1,""]},"autopilot.core.networking.Terminal_Station":{l_change:[2,2,1,""],l_continuous:[2,2,1,""],l_data:[2,2,1,""],l_file:[2,2,1,""],l_handshake:[2,2,1,""],l_init:[2,2,1,""],l_kill:[2,2,1,""],l_ping:[2,2,1,""],l_state:[2,2,1,""],l_stopall:[2,2,1,""],plot_timer:[2,3,1,""],sent_plot:[2,3,1,""],start_plot_timer:[2,2,1,""]},"autopilot.agents.pilot":{Pilot:[3,1,1,""]},"autopilot.agents.pilot.Pilot":{blank_LEDs:[3,2,1,""],calibrate_port:[3,2,1,""],calibration_curve:[3,2,1,""],file_block:[3,3,1,""],get_ip:[3,2,1,""],handshake:[3,2,1,""],init_audio:[3,2,1,""],init_logging:[3,2,1,""],init_pigpio:[3,2,1,""],l_bandwidth:[3,2,1,""],l_cal_port:[3,2,1,""],l_cal_result:[3,2,1,""],l_param:[3,2,1,""],l_start:[3,2,1,""],l_stop:[3,2,1,""],log_formatter:[3,3,1,""],log_handler:[3,3,1,""],logger:[3,3,1,""],networking:[3,3,1,""],node:[3,3,1,""],open_file:[3,2,1,""],run_task:[3,2,1,""],running:[3,3,1,""],server:[3,3,1,""],stage_block:[3,3,1,""],update_state:[3,2,1,""]},"autopilot.core.subject":{Subject:[6,1,1,""]},"autopilot.core.subject.Subject":{Hash_Table:[6,1,1,""],History_Table:[6,1,1,""],Weight_Table:[6,1,1,""],apply_along:[6,2,1,""],assign_protocol:[6,2,1,""],close_hdf:[6,2,1,""],data_thread:[6,2,1,""],ensure_structure:[6,2,1,""],flush_current:[6,2,1,""],get_step_history:[6,2,1,""],get_timestamp:[6,2,1,""],get_trial_data:[6,2,1,""],get_weight:[6,2,1,""],graduate:[6,2,1,""],new_subject_file:[6,2,1,""],open_hdf:[6,2,1,""],prepare_run:[6,2,1,""],save_data:[6,2,1,""],set_weight:[6,2,1,""],stash_current:[6,2,1,""],stop_run:[6,2,1,""],to_csv:[6,2,1,""],update_biography:[6,2,1,""],update_history:[6,2,1,""],update_weights:[6,2,1,""]},"autopilot.core.subject.Subject.Hash_Table":{columns:[6,3,1,""]},"autopilot.core.subject.Subject.History_Table":{columns:[6,3,1,""]},"autopilot.core.subject.Subject.Weight_Table":{columns:[6,3,1,""]},"autopilot.core.utils":{Param:[8,1,1,""],ReturnThread:[8,1,1,""],coerce_discrete:[8,4,1,""],list_subjects:[8,4,1,""],load_pilotdb:[8,4,1,""]},"autopilot.core.utils.Param":{__contains__:[8,2,1,""],__delitem__:[8,2,1,""],__getitem__:[8,2,1,""],__setitem__:[8,2,1,""],tag:[8,3,1,""],type:[8,3,1,""],types:[8,3,1,""]},"autopilot.core.utils.ReturnThread":{join:[8,2,1,""],run:[8,2,1,""]},"autopilot.hardware":{BCM_TO_BOARD:[10,5,1,""],BOARD_TO_BCM:[10,5,1,""],Hardware:[10,1,1,""],cameras:[11,0,0,"-"],gpio:[12,0,0,"-"],i2c:[13,0,0,"-"],usb:[14,0,0,"-"]},"autopilot.hardware.Hardware":{assign_cb:[10,2,1,""],get_name:[10,2,1,""],init_logging:[10,2,1,""],init_networking:[10,2,1,""],input:[10,3,1,""],is_trigger:[10,3,1,""],output:[10,3,1,""],pin:[10,3,1,""],release:[10,2,1,""],type:[10,3,1,""]},"autopilot.hardware.cameras":{Camera:[11,1,1,""],Camera_CV:[11,1,1,""],Camera_Spinnaker:[11,1,1,""],OPENCV_LAST_INIT_TIME:[11,5,1,""],Video_Writer:[11,1,1,""],list_spinnaker_cameras:[11,4,1,""]},"autopilot.hardware.cameras.Camera":{_capture:[11,2,1,""],_grab:[11,2,1,""],_process:[11,2,1,""],_timestamp:[11,2,1,""],_write_deinit:[11,2,1,""],_write_frame:[11,2,1,""],cam:[11,2,1,""],capture:[11,2,1,""],capture_deinit:[11,2,1,""],capture_init:[11,2,1,""],init_cam:[11,2,1,""],input:[11,3,1,""],l_start:[11,2,1,""],l_stop:[11,2,1,""],output_filename:[11,2,1,""],queue:[11,2,1,""],release:[11,2,1,""],stop:[11,2,1,""],stream:[11,2,1,""],type:[11,3,1,""],write:[11,2,1,""]},"autopilot.hardware.cameras.Camera_CV":{_grab:[11,2,1,""],_timestamp:[11,2,1,""],backend:[11,2,1,""],fps:[11,2,1,""],init_cam:[11,2,1,""],release:[11,2,1,""],shape:[11,2,1,""],v4l_info:[11,2,1,""]},"autopilot.hardware.cameras.Camera_Spinnaker":{ATTR_TYPES:[11,3,1,""],ATTR_TYPE_NAMES:[11,3,1,""],RW_MODES:[11,3,1,""],_grab:[11,2,1,""],_process:[11,2,1,""],_timestamp:[11,2,1,""],_write_deinit:[11,2,1,""],_write_frame:[11,2,1,""],acquisition_mode:[11,2,1,""],bin:[11,2,1,""],capture_deinit:[11,2,1,""],capture_init:[11,2,1,""],device_info:[11,2,1,""],exposure:[11,2,1,""],fps:[11,2,1,""],frame_trigger:[11,2,1,""],get:[11,2,1,""],init_cam:[11,2,1,""],list_options:[11,2,1,""],readable_attributes:[11,2,1,""],release:[11,2,1,""],set:[11,2,1,""],writable_attributes:[11,2,1,""],write:[11,2,1,""]},"autopilot.hardware.cameras.Video_Writer":{run:[11,2,1,""]},"autopilot.hardware.gpio":{Digital_In:[12,1,1,""],Digital_Out:[12,1,1,""],ENABLED:[12,5,1,""],GPIO:[12,1,1,""],INVERSE_PULL_MAP:[12,5,1,""],INVERSE_TRIGGER_MAP:[12,5,1,""],LED_RGB:[12,1,1,""],PULL_MAP:[12,5,1,""],PWM:[12,1,1,""],Solenoid:[12,1,1,""],TRIGGER_MAP:[12,5,1,""]},"autopilot.hardware.gpio.Digital_In":{assign_cb:[12,2,1,""],clear_cb:[12,2,1,""],input:[12,3,1,""],is_trigger:[12,3,1,""],record_event:[12,2,1,""],release:[12,2,1,""],type:[12,3,1,""]},"autopilot.hardware.gpio.Digital_Out":{_series_script:[12,2,1,""],delete_script:[12,2,1,""],output:[12,3,1,""],pigs_function:[12,3,1,""],pulse:[12,2,1,""],release:[12,2,1,""],series:[12,2,1,""],set:[12,2,1,""],stop_script:[12,2,1,""],store_series:[12,2,1,""],toggle:[12,2,1,""],turn:[12,2,1,""],type:[12,3,1,""]},"autopilot.hardware.gpio.GPIO":{init_pigpio:[12,2,1,""],pin:[12,2,1,""],polarity:[12,2,1,""],pull:[12,2,1,""],release:[12,2,1,""],trigger:[12,2,1,""]},"autopilot.hardware.gpio.LED_RGB":{_series_script:[12,2,1,""],flash:[12,2,1,""],output:[12,3,1,""],pin:[12,2,1,""],pin_bcm:[12,2,1,""],pull:[12,2,1,""],pulse:[12,2,1,""],range:[12,2,1,""],release:[12,2,1,""],set:[12,2,1,""],toggle:[12,2,1,""],type:[12,3,1,""]},"autopilot.hardware.gpio.PWM":{output:[12,3,1,""],pigs_function:[12,3,1,""],polarity:[12,2,1,""],range:[12,2,1,""],release:[12,2,1,""],set:[12,2,1,""],type:[12,3,1,""]},"autopilot.hardware.gpio.Solenoid":{DURATION_MIN:[12,3,1,""],dur_from_vol:[12,2,1,""],duration:[12,2,1,""],open:[12,2,1,""],output:[12,3,1,""],type:[12,3,1,""]},"autopilot.hardware.i2c":{I2C_9DOF:[13,1,1,""],MLX90640:[13,1,1,""]},"autopilot.hardware.i2c.I2C_9DOF":{ACCELRANGE_16G:[13,3,1,""],ACCELRANGE_2G:[13,3,1,""],ACCELRANGE_4G:[13,3,1,""],ACCELRANGE_8G:[13,3,1,""],GYROSCALE_2000DPS:[13,3,1,""],GYROSCALE_245DPS:[13,3,1,""],GYROSCALE_500DPS:[13,3,1,""],MAGGAIN_12GAUSS:[13,3,1,""],MAGGAIN_16GAUSS:[13,3,1,""],MAGGAIN_4GAUSS:[13,3,1,""],MAGGAIN_8GAUSS:[13,3,1,""],accel_range:[13,2,1,""],acceleration:[13,2,1,""],gyro:[13,2,1,""],gyro_scale:[13,2,1,""],mag_gain:[13,2,1,""],magnetic:[13,2,1,""],temperature:[13,2,1,""]},"autopilot.hardware.i2c.MLX90640":{ALLOWED_FPS:[13,3,1,""],SHAPE_SENSOR:[13,3,1,""],_grab:[13,2,1,""],_threaded_capture:[13,2,1,""],_timestamp:[13,2,1,""],capture_init:[13,2,1,""],fps:[13,2,1,""],init_cam:[13,2,1,""],integrate_frames:[13,2,1,""],interpolate:[13,2,1,""],interpolate_frame:[13,2,1,""],release:[13,2,1,""],type:[13,3,1,""]},"autopilot.hardware.usb":{Scale:[14,1,1,""],Wheel:[14,1,1,""]},"autopilot.hardware.usb.Scale":{MODEL:[14,3,1,""]},"autopilot.hardware.usb.Wheel":{MODES:[14,3,1,""],MOVE_DTYPE:[14,3,1,""],THRESH_TYPES:[14,3,1,""],assign_cb:[14,2,1,""],calc_move:[14,2,1,""],check_thresh:[14,2,1,""],input:[14,3,1,""],l_clear:[14,2,1,""],l_measure:[14,2,1,""],l_stop:[14,2,1,""],release:[14,2,1,""],start:[14,2,1,""],thresh_trig:[14,2,1,""],trigger:[14,3,1,""],type:[14,3,1,""]},"autopilot.prefs":{add:[15,4,1,""],compute_calibration:[15,4,1,""],git_version:[15,4,1,""],init:[15,4,1,""],prefdict:[15,5,1,""]},"autopilot.setup":{setup_pilot:[17,0,0,"-"],setup_scale:[18,0,0,"-"],setup_terminal:[19,0,0,"-"]},"autopilot.setup.setup_pilot":{HardwareForm:[17,1,1,""],PilotSetupForm:[17,1,1,""],SetupApp:[17,1,1,""],make_dir:[17,4,1,""],unfold_values:[17,4,1,""]},"autopilot.setup.setup_pilot.HardwareForm":{afterEditing:[17,2,1,""],create:[17,2,1,""]},"autopilot.setup.setup_pilot.PilotSetupForm":{afterEditing:[17,2,1,""],create:[17,2,1,""]},"autopilot.setup.setup_pilot.SetupApp":{onStart:[17,2,1,""]},"autopilot.setup.setup_terminal":{SetupApp:[19,1,1,""],TerminalSetupForm:[19,1,1,""],make_dir:[19,4,1,""],unfold_values:[19,4,1,""]},"autopilot.setup.setup_terminal.SetupApp":{onStart:[19,2,1,""]},"autopilot.setup.setup_terminal.TerminalSetupForm":{afterEditing:[19,2,1,""],create:[19,2,1,""]},"autopilot.stim":{managers:[21,0,0,"-"],sound:[22,0,0,"-"]},"autopilot.stim.managers":{Bias_Correction:[21,1,1,""],Proportional:[21,1,1,""],Stim_Manager:[21,1,1,""],init_manager:[21,4,1,""]},"autopilot.stim.managers.Bias_Correction":{next_bias:[21,2,1,""],thresholded_linear:[21,2,1,""],update:[21,2,1,""]},"autopilot.stim.managers.Proportional":{init_sounds_grouped:[21,2,1,""],init_sounds_individual:[21,2,1,""],next_stim:[21,2,1,""],set_triggers:[21,2,1,""],store_groups:[21,2,1,""]},"autopilot.stim.managers.Stim_Manager":{compute_correction:[21,2,1,""],do_bias:[21,2,1,""],do_correction:[21,2,1,""],end:[21,2,1,""],init_sounds:[21,2,1,""],make_punishment:[21,2,1,""],next_stim:[21,2,1,""],play_punishment:[21,2,1,""],set_triggers:[21,2,1,""],update:[21,2,1,""]},"autopilot.stim.sound":{jackclient:[23,0,0,"-"],pyoserver:[24,0,0,"-"],sounds:[25,0,0,"-"]},"autopilot.stim.sound.jackclient":{BLOCKSIZE:[23,5,1,""],CONTINUOUS:[23,5,1,""],CONTINUOUS_LOOP:[23,5,1,""],CONTINUOUS_QUEUE:[23,5,1,""],FS:[23,5,1,""],JackClient:[23,1,1,""],PLAY:[23,5,1,""],QUEUE:[23,5,1,""],Q_LOCK:[23,5,1,""],SERVER:[23,5,1,""],STOP:[23,5,1,""]},"autopilot.stim.sound.jackclient.JackClient":{boot_server:[23,2,1,""],process:[23,2,1,""],quit:[23,2,1,""],run:[23,2,1,""]},"autopilot.stim.sound.pyoserver":{pyo_server:[24,4,1,""]},"autopilot.stim.sound.sounds":{File:[25,1,1,""],Gap:[25,1,1,""],Noise:[25,1,1,""],SOUND_LIST:[25,5,1,""],STRING_PARAMS:[25,5,1,""],Speech:[25,1,1,""],Tone:[25,1,1,""],int_to_float:[25,4,1,""]},"autopilot.stim.sound.sounds.File":{PARAMS:[25,3,1,""],init_sound:[25,2,1,""],type:[25,3,1,""]},"autopilot.stim.sound.sounds.Gap":{PARAMS:[25,3,1,""],buffer:[25,2,1,""],chunk:[25,2,1,""],init_sound:[25,2,1,""],play:[25,2,1,""],type:[25,3,1,""]},"autopilot.stim.sound.sounds.Noise":{PARAMS:[25,3,1,""],init_sound:[25,2,1,""],type:[25,3,1,""]},"autopilot.stim.sound.sounds.Speech":{PARAMS:[25,3,1,""],type:[25,3,1,""]},"autopilot.stim.sound.sounds.Tone":{PARAMS:[25,3,1,""],init_sound:[25,2,1,""],type:[25,3,1,""]},"autopilot.tasks":{CHILDREN_LIST:[26,5,1,""],TASK_LIST:[26,5,1,""],children:[27,0,0,"-"],free_water:[28,0,0,"-"],graduation:[29,0,0,"-"],nafc:[30,0,0,"-"],task:[31,0,0,"-"]},"autopilot.tasks.children":{Transformer:[27,1,1,""],Video_Child:[27,1,1,""],Wheel_Child:[27,1,1,""]},"autopilot.tasks.children.Transformer":{l_process:[27,2,1,""],noop:[27,2,1,""]},"autopilot.tasks.children.Video_Child":{PARAMS:[27,3,1,""],noop:[27,2,1,""],start:[27,2,1,""],stop:[27,2,1,""]},"autopilot.tasks.children.Wheel_Child":{HARDWARE:[27,3,1,""],PARAMS:[27,3,1,""],STAGE_NAMES:[27,3,1,""],end:[27,2,1,""],noop:[27,2,1,""]},"autopilot.tasks.free_water":{Free_Water:[28,1,1,""]},"autopilot.tasks.free_water.Free_Water":{DATA:[28,3,1,""],HARDWARE:[28,3,1,""],PARAMS:[28,3,1,""],PLOT:[28,3,1,""],STAGE_NAMES:[28,3,1,""],TrialData:[28,1,1,""],end:[28,2,1,""],response:[28,2,1,""],water:[28,2,1,""]},"autopilot.tasks.free_water.Free_Water.TrialData":{columns:[28,3,1,""]},"autopilot.tasks.graduation":{Accuracy:[29,1,1,""],GRAD_LIST:[29,5,1,""],Graduation:[29,1,1,""],NTrials:[29,1,1,""]},"autopilot.tasks.graduation.Accuracy":{COLS:[29,3,1,""],PARAMS:[29,3,1,""],update:[29,2,1,""]},"autopilot.tasks.graduation.Graduation":{COLS:[29,3,1,""],PARAMS:[29,3,1,""],update:[29,2,1,""]},"autopilot.tasks.graduation.NTrials":{PARAMS:[29,3,1,""],update:[29,2,1,""]},"autopilot.tasks.nafc":{Nafc:[30,1,1,""],Nafc_Gap:[30,1,1,""]},"autopilot.tasks.nafc.Nafc":{HARDWARE:[30,3,1,""],PARAMS:[30,3,1,""],PLOT:[30,3,1,""],STAGE_NAMES:[30,3,1,""],TrialData:[30,1,1,""],discrim:[30,2,1,""],flash_leds:[30,2,1,""],punish:[30,2,1,""],reinforcement:[30,2,1,""],request:[30,2,1,""],respond:[30,2,1,""],stim_end:[30,2,1,""],stim_start:[30,2,1,""]},"autopilot.tasks.nafc.Nafc.TrialData":{columns:[30,3,1,""]},"autopilot.tasks.nafc.Nafc_Gap":{PARAMS:[30,3,1,""],end:[30,2,1,""]},"autopilot.tasks.task":{Task:[31,1,1,""]},"autopilot.tasks.task.Task":{HARDWARE:[31,3,1,""],PARAMS:[31,3,1,""],PLOT:[31,3,1,""],STAGE_NAMES:[31,3,1,""],TrialData:[31,3,1,""],end:[31,2,1,""],flash_leds:[31,2,1,""],handle_trigger:[31,2,1,""],init_hardware:[31,2,1,""],set_leds:[31,2,1,""],set_reward:[31,2,1,""]},"autopilot.viz":{psychometric:[33,0,0,"-"],trial_viewer:[34,0,0,"-"]},"autopilot.viz.psychometric":{calc_psychometric:[33,4,1,""],plot_psychometric:[33,4,1,""]},"autopilot.viz.trial_viewer":{load_subject_data:[34,4,1,""],load_subject_dir:[34,4,1,""],step_viewer:[34,4,1,""],trial_viewer:[34,4,1,""]},autopilot:{core:[0,0,0,"-"],hardware:[10,0,0,"-"],prefs:[15,0,0,"-"],setup:[16,0,0,"-"],stim:[20,0,0,"-"],tasks:[26,0,0,"-"],viz:[32,0,0,"-"]}},objnames:{"0":["py","module","Python module"],"1":["py","class","Python class"],"2":["py","method","Python method"],"3":["py","attribute","Python attribute"],"4":["py","function","Python function"],"5":["py","data","Python data"]},objtypes:{"0":"py:module","1":"py:class","2":"py:method","3":"py:attribute","4":"py:function","5":"py:data"},terms:{"01t17":6,"100u":12,"1mhz":[13,37],"2000m":40,"20m":44,"2afc":[8,26,44],"2afc_gap":26,"30fp":11,"30khz":46,"32x24px":37,"3rd":[35,46],"400khz":13,"50m":11,"64fp":[9,13,37],"807693v1":[36,42],"8mb":40,"9dof":[13,37],"abstract":[38,46],"boolean":43,"break":[2,44],"byte":12,"case":[8,11,12,23,37,38,46],"class":[2,3,6,8,10,11,12,13,14,17,19,21,23,25,26,27,28,29,30,31,37,39,42,44,45],"default":[2,3,6,8,11,12,13,17,19,23,37,38,40,41,43,44,46],"enum":46,"export":[6,37],"final":[25,37,43,44],"float":[3,6,8,11,12,13,21,25,29,30,31,38,43],"function":[1,2,4,7,8,10,11,12,14,15,17,19,21,24,25,31,33,34,37,38,40,43,46],"import":[12,15,37,38,42,43,46],"int":[2,3,6,8,10,11,12,13,14,21,23,25,27,28,29,30,31,38,43],"long":[3,6,10,46],"new":[2,3,6,11,21,35,43,44,46],"return":[2,3,6,8,9,10,11,12,13,14,15,21,24,25,27,28,29,30,31,33,34,37,43],"short":[42,43,46],"static":40,"super":43,"switch":[37,39,44,46],"true":[2,3,6,10,11,12,14,21,24,25,27,29,30,31,34,38,43,44,46],"try":[2,8,11,38,43],"while":[6,11,30,37,40,43,46],Added:37,Adding:44,And:[15,40,43],Are:[12,21,40],FPS:[11,13],For:[11,12,21,35,38,42,44,45,46],IDs:12,Its:[38,45],NOT:[15,23],Not:[2,3,8,10,11,14,18,21,38],One:[11,21],Pis:[42,44,45],That:[37,44,46],The:[2,3,6,8,10,11,12,13,17,21,23,25,27,28,37,38,39,40,41,42,44,46],Then:[2,43],There:[2,38,43,46],These:[2,12,25,39,40,43,44],Use:[11,24,31,40,46],Used:[2,3,6,12,43],Uses:[12,13,14],Using:11,Was:21,With:[30,38,46],Yes:[6,44],__contains__:[2,8],__del__:[10,38],__delitem__:[2,8],__dict__:2,__getitem__:[2,8],__init__:[8,10,38,43,46],__setitem__:[2,8],_cam:11,_camera_method:11,_captur:11,_clean_valu:12,_deserialize_numpi:37,_frame:13,_grab:[11,13],_grab_ev:13,_mous:43,_output_filenam:11,_process:11,_record:43,_serialize_numpi:[2,37],_series_script:12,_step_nam:6,_stream:2,_threaded_captur:13,_timestamp:[11,13],_write_deinit:11,_write_fram:11,_write_q:11,abbrevi:43,abil:[6,37,46],abl:[10,46],about:[6,11,25,37,46],abov:[14,21,23,29,40,43,44],absolut:[11,37],abyz:12,accel:13,accel_rang:13,acceler:[13,46],acceleromet:[13,37,46],accelrange_16g:13,accelrange_2g:13,accelrange_4g:13,accelrange_8g:13,accept:[10,11,12,14],access:[6,23,37,40,43,44],accord:[2,3,6,11,13,21,25,37,42],accordingli:[2,12,13,38,46],accur:[37,43,46],accuraci:[29,42,44],achiev:37,acquir:[11,37,46],acquisit:[11,13,46],acquisition_mod:11,acquisitionframer:11,across:[13,37,40,43,46],activ:[8,10,11,14,23,30,43],actual:[38,42,44],adafruit:13,adapt:46,add:[6,12,15,21,38,40,43,44,46],added:[2,25,37,40,44,46],adding:[12,44],addit:[10,12,37,40,42,46],addition:[43,46],address:[2,3,17,40,43,44],addus:40,adir:[17,19],admin:40,adopt:46,advanc:[40,43],advantag:11,advis:44,aesthet:37,affect:38,after:[2,6,8,11,12,15,23,29,34,37,38,40,41,43,44],afteredit:[17,19],afterward:28,agent:[15,27,37,39,41,42,46],alert:[2,28,46],algorithm:46,align:[43,46],aliv:[2,3,8,12,23],all:[2,6,10,11,12,19,21,25,28,29,31,37,38,39,40,41,43,44,46],allcap:15,alloc:[11,40],allow:[11,12,28,37,38,40,42,43,44,45,46],allow_repeat:[28,44],allowed_fp:13,almost:45,along:[2,6,25,43],alongsid:[31,46],alreadi:[2,6,11,12,40,46],alsa:40,alsafil:40,also:[8,10,12,37,38,40,43,45,46],altair:[33,37],altern:[6,30,44],alwai:[2,8,10,11,13,37,40,46],ambit:37,amount:[3,12,23],amp:40,amplifi:40,amplitud:[25,30,44],analog:42,analysi:46,analyz:46,ancient:46,angl:43,ani:[2,3,8,10,11,12,13,14,15,17,19,28,29,31,38,39,40,43,44,45,46],anim:44,anod:[12,38],anoth:[2,11,12,37,38,43,46],answer:[21,30,44],anyth:[6,8,10],anywai:[6,11],api:[38,42,45,46],aplai:40,apnpsapplicationmanag:[17,19],apparatu:42,appear:44,append:[6,11,38],appli:[12,38,46],applic:[11,38,46],apply_along:6,appropri:[2,6,15,37,43,44],apt:40,arab:40,arbitrari:[21,42,45,46],architectur:[37,46],aren:37,arena:[37,44,46],arg:[2,8,11,13,17,19,21,28,30,31,43],argument:[8,10,11,12,28,43],aris:6,arm:40,around:[11,12,37,43,46],arrai:[2,13,14,23,25,37,38,43,45,46],arrang:37,asctim:[2,3],ask:[2,3,17,43],asound:40,assembl:46,assign:[6,12,31,43,44],assign_cb:[10,12,14,38],assign_protocol:6,assum:[12,38,43,46],astunpars:40,asynchron:30,attempt:[2,8,11,37,38,44],attr:[11,12,13],attr_typ:11,attr_type_nam:11,attribut:[2,6,8,10,11,12,13,14,25,27,28,29,30,31,37,38],attributeerror:43,audibl:40,audio:[3,17,23,24,25,37,39,46],audioserv:[3,15,17,25],auditori:46,auto:[11,46],autom:[41,42,46],automat:[11,37,44,45,46],autopilot:[2,3,5,9,11,12,13,14,15,17,19,21,25,26,27,28,29,30,36,37,38,39,40,41,42,43,44,45],autopilot_root:15,autpilot:35,avail:[11,12,37,40,43,46],averag:[11,13,37],avoid:[11,25,37,40,46],await:13,awar:44,awkward:[37,46],axi:[13,43],b64:2,b64encod:2,back:[2,3,31,43],backend:[11,40],background:[12,23,38],bad:[6,10,43],bail:[30,31,43],balanc:37,bar:[31,34,43,46],base64:2,base:[2,3,6,8,10,11,12,13,14,17,19,21,23,25,27,28,29,30,31,37,38,40,41,44,46],base_dir:[40,41],base_path:11,basedir:[3,15,17,19,41,44],baselin:6,baseprocess:11,basestr:6,bash:[10,46],basic:[21,37,38,40,41,43,44,46],batch:[37,46],baudrat:[9,13,37],bcm:[10,12,31,38],bcm_to_board:10,beam:44,beambreak:[3,17,31,40,43],beauti:46,becaus:[11,37,43,44],becom:46,been:[2,8,11,12,25,37,38,39,40,43,45,46],befor:[6,8,11,15,23,30,37,38,40,43,46],begin:[11,37,43,44],beginacquisit:11,begun:37,behav:[37,43],behavior:[2,42,45],behind:2,beholden:46,being:[2,6,11,12,37,38,43,46],bell:38,below:[14,41,43,44],better:[25,37,40,46],between:[2,3,10,21,26,28,29,37,39,40,42,43,46],bia:[21,30,43],bias:21,bias_correct:[21,30],bias_mod:[30,43,44],bias_threshold:[30,43],bidi:40,bin:[11,15,46],binari:[3,11,17,38],bind:[2,40],biograph:[6,44],biographi:6,biography_tab:6,biorxiv:[36,42],bit:[13,25,37,38,43,46],bittorr:46,blank:[2,43],blank_l:3,blend:46,blink:[12,38],block:[3,8,11,38,43],blocksiz:23,blosc:[2,11,37,40,41],blue:[12,30,43],bluealsa:40,bluetooth:40,board:[3,10,12,17,37,38,45],board_to_bcm:[10,38],bodi:2,bom:46,bool:[2,6,8,10,11,12,14,21,24,25,28,29,30,31,34,38,43],boolcol:43,boost:37,boot:[3,13,17,23,24,31,40],boot_serv:23,border:46,both:[12,38,40,43,46],both_edg:12,bother:25,bottleneck:46,bound:2,brand:11,brandi:46,breviti:38,brew:41,bridg:10,bright:46,bring:12,broadcom:[10,12],brought:[37,46],brows:46,buffer:[11,13,23,25,37,40,43],bug:37,bugfix:35,build:[37,40,43,44,46],build_pi:46,built:[21,38,42,46],bunch:37,button:[44,46],c3c7dcb32f164e2740b4fca98237908e74a00a4c:15,c_doubl:11,cach:[3,23,43],cad:46,cal_progress:3,calc_mov:14,calc_psychometr:33,calcul:[14,21,33,37,43],calibr:[3,12,13,15,25,37,41,46],calibrate_port:[3,12],calibration_curv:3,call:[2,3,6,8,10,11,12,13,14,17,19,21,23,25,30,31,33,37,38,39,40,43,44,46],callabl:[8,12,21,28,43],callback:[9,10,12,21,30,31,38,43,46],callback_fn:[12,38],cam:[11,27],cam_list:11,camera:[10,13,27,37,45,46],camera_cv:[11,37],camera_idx:11,camera_spinnak:[11,37],cameralist:11,can:[2,6,8,10,11,12,13,21,23,25,30,36,37,38,40,42,43,44,45,46],cancel:[12,43],cannot:12,canon:12,cap_prop_fp:11,cap_prop_frame_width:11,cap_prop_pos_msec:11,capabl:[9,13,37,46],captur:[9,11,13,37,43,45],capture_:11,capture_deinit:11,capture_directori:11,capture_init:[11,13],card:40,care:[10,38,40],carri:28,cascad:46,categori:44,cathod:[12,38],caus:[6,8,23,40,46],cautiou:11,ceaseless:37,center:[11,30,43,44],certain:43,cffi:40,cfg:40,challeng:44,chanc:21,chance_bar:[30,31,43],chance_level:43,chang:[2,3,6,12,25,27,36,37,40,41,43,44,46],change_to_blu:43,changelog:45,chann:12,channel:[3,11,12,17,23,40],charact:[30,43],chart:33,check:[2,3,6,11,14,37,40,43,45,46],check_thresh:[14,43],checkbox:43,child:[2,26,27,37,40,46],childid:43,children:[2,26,42,43,46],children_list:[26,43],chip:10,choic:[21,30,44],choos:[21,40,43,46],chosen:46,chunk:25,circuitfun:13,circular:43,clariti:38,classic:46,clean:[13,38],cleanli:46,cleanup:38,clear:[3,12,14,23,38,43],clear_cb:12,click:[38,44,46],client:[10,23,40],clock:[40,46],clone:40,close:[6,12,37],close_hdf:6,clumsi:46,cmake:40,coars:6,code:[13,35,38,43,46],coeffici:12,coerc:8,coerce_discret:8,coher:[2,3,8],col:[8,29,38,43],col_nam:6,collabor:46,collect:[21,27,31,34,37,38,43,44,46],collis:46,color:[3,12,17,31,38,40,43],color_dict:31,color_seri:38,column:[6,8,25,28,29,30,33,37,43,44,46],com:[8,11,12,13,14,35,37,40],combin:[12,13,25,37,42,45],come:[25,37,43,44],command:[10,17,19,40,44,46],comment:40,commit:[6,15,35],common:[12,37,38,44,46],commonli:[12,43],commun:[2,3,38,42,45,46],compat:[6,37],compil:[11,40],complet:[11,12,30,31,37,42,43],complex:[2,43,45,46],compon:[37,42,45],comprehens:42,compres:11,compress:[2,11,37,40,45,46],comput:[2,3,12,14,21,30,37,38,42,43,46],compute_calibr:15,compute_correct:21,concept:43,concurr:46,condit:[27,46],conf:40,config:[13,15,37,40],configobj:40,configur:[3,11,17,37,38,40,41,42,43,44,46],confin:43,confirm:[2,37],conflict:37,confus:[37,45],confusingli:40,connect:[2,3,12,13,17,23,37,38,40,43,46],consid:[29,46],consist:[36,42,44],consol:46,conson:25,constant:38,constitut:42,construct:[8,12,42,43],constructor:8,consum:[11,44,46],contain:[2,28,37,46],content:[36,42],context:[2,11,23],continu:[2,3,11,13,14,21,23,25,30,31,37,43,46],continuous_cycl:23,continuous_data:6,continuous_loop:23,continuous_queu:23,continuousdata:[43,46],contribut:43,control:[2,3,11,17,37,40,41,42,43,44,46],conveni:[12,38],convers:11,convert:[10,11,12,25,37,38,46],cookbook:46,cool_new_step:44,coordin:[3,42,45,46],copi:[2,40],core:[2,37,43,45],correct:[21,28,29,30,31,37,40,43,44],correction_pct:[21,30,43,44],correction_tri:[21,30,43,46],correctli:[9,44],correspond:[12,29,40,44,46],corrupt:6,costli:46,could:[37,38,40,46],count:[2,12,28,29,30,43],counter:[2,29],cpu:40,cpufreq:40,creat:[2,3,6,11,12,17,19,25,37,38,39,40,41,43,45,46],creation:[2,46],credenti:40,credit:40,crew:46,criteria:[29,42,44],critic:[37,40],cross:43,crude:46,css:37,csv:[6,11],ctl:[11,40],ctrl:40,ctx:2,cumbersom:46,current:[2,3,6,8,10,11,12,15,17,21,28,30,37,38,39,40,41,42,43,44,46],current_stag:[30,43],current_tri:[6,28,29,30,43,44],curv:[3,33,37],custom:[12,25,40],customiz:42,cute:12,cv2:11,cycl:[11,12,23,28,37,38,43,46],cycle_widget:17,dac:40,dacplu:40,daemon:[2,10,12,37,38,43,44],dag:46,dai:3,dalsa:[3,17,40],data:[2,3,6,10,11,12,15,19,23,25,26,28,29,30,31,33,34,37,40,41,42,44,45,46],data_dir:34,data_queu:6,data_thread:6,databas:46,datadir:[3,6,15,19,41],datafram:[6,8,33],date:6,date_protocol_nam:6,datetim:[2,13,30,43],dc_timestamp:[30,31,43],deadlock:8,dealer:[2,46],debug:[11,15,24,37,43],decid:[8,21,37],decim:8,decis:37,declar:37,decod:[2,46],decompress:2,decor:46,decreas:40,decrement:2,dedic:37,def:[38,43],defin:[2,3,6,10,17,25,29,31,38,42,43],degre:13,delai:[38,43],delayed_set:43,delet:[6,12,38],delete_script:12,delin:43,deliv:[30,43,45],deliveri:[12,37,43,45],demand:[10,40],demonstr:43,denot:43,dep:40,depend:[3,10,11,14,21,25,30,37,40,41,43,46],deprec:40,depth:[37,40,43],dequ:21,deriv:[17,19],descend:46,describ:[6,10,21,37,43,44,46],descript:[2,6,11,12,26,28,30,31,37,38,42,43,46],descriptor:[11,37,43],deseri:37,design:[10,27,37,38,43,45,46],desir:[3,11,12],desktop:42,destroi:37,detail:[6,11,35,37,42,43,45],detect:[12,38,44],determin:[2,23,46],dev:40,dev_debug:11,develop:[10,37,39,40,41,42,43,45,46],devic:[2,10,11,12,13,14,37,38,40,43,46],device_info:11,deviceinform:11,dflt:[6,28,30],dhw:[3,17,40],dialog:37,dict:[2,3,6,8,10,11,12,15,17,19,21,27,28,30,31,38,43],dictionari:[2,3,6,8,10,11,15,21,27,28,31,37,40,43,44],did:[29,46],did_gradu:6,didn:[6,37],die:2,differ:[2,11,13,21,29,37,38,39,42,43,46],differenti:37,difficult:46,difficulti:11,digi_in:12,digi_out:[14,43],digit:[12,14,18,37,38,40],digital_in:[12,28,30,37,38,40],digital_out:[12,14,27,37,43],dir:[6,14],direct:[2,12,21,38,40,46],directli:[2,11,23,40,43,46],directori:[3,11,15,17,19,34,40,41,44,46],directory_writ:11,disabl:[40,44],discard:23,discourag:24,discret:[8,10,37],discrim:30,discrim_plai:[30,43],discrimin:44,discriminand:[30,33],discuss:[37,38,43,45],dish:46,disk2:40,disk:[11,40,43],diskutil:40,disorgan:37,dispens:[28,40],dispers:3,displai:[37,44,46],dissect:38,dist:[14,43],distanc:[14,27,43],distinct:11,distinctli:46,distractor:[21,30,43],distribut:[2,42,45,46],disturb:46,distutil:40,divid:[12,38],dlc:46,dlc_latenc:26,dmesg:11,do_bia:[21,43],do_correct:[21,43],do_log:37,do_return:15,do_trigg:43,dob:6,doc:[3,5,6,10,12,15,17,37,46],docker:46,document:[11,38,40,42,43,45],doe300:46,doe:[2,6,11,12,25,39,40,41,43,44],doesn:[11,12,17,19,43,44,46],doing:[8,21,25,29],don:[2,37,38,40,42,43,46],done:46,dos:46,doubl:[2,46],down:[3,12,17,28,37,38,43],download:[40,46],download_box:37,dpkg:40,draft:21,drag:46,dramat:[37,41],draw:[31,43],draw_line_at:19,drawfp:[2,15,19,37,41],drawn:43,drift:43,drive:3,driver:[9,37,40],drop:[2,23,46],dropout:24,drown:37,dtoverlai:40,dtparam:[13,40],dtype:38,due:24,dump:[2,6,11,23,37,40],duplic:[37,46],dur_from_vol:12,durat:[3,11,12,21,25,28,30,31,38,43,44],duration_min:12,dure:[11,21,31],duti:[12,38],dutycycl:12,each:[2,3,6,10,11,12,17,23,25,27,30,31,33,34,37,38,39,41,42,43,44,46],earli:43,earlier:44,easier:[10,37,39,45,46],easili:37,echo:[11,40],edg:[12,38],edit:[13,40,44,46],effect:[23,25,42,43],effici:[45,46],effort:[37,46],eft:[43,44],egregi:43,either:[2,3,6,8,11,12,27,37,38,40,43,44,46],either_edg:12,elaps:43,element:[43,46],els:[6,8,12,25,29,43],elsewher:25,elud:46,embarassingli:37,empti:[11,23,38],enabl:[11,12,37,40,44,46],encapsul:2,encod:[2,11,37],end:[6,11,12,21,23,27,28,30,31,37,43,44],end_thread:38,endless:43,enforc:[23,38],engin:37,enhanc:11,enough:37,ensur:[6,43,46],ensure_structur:6,enter:43,entir:[40,46],entri:[2,3,40,43,44],entropi:37,environ:[39,46],equal:38,equival:6,error:[8,11,37,45,46],especi:38,esprima:40,essenti:[39,40],establish:[3,38],etc:[2,6,10,11,12,37,40,43,44,46],eter:46,even:[10,37],event:[2,3,6,10,11,12,13,23,28,30,31,37,38,43,45,46],everi:[2,3,6,10,13,14,37,44,45,46],everyon:2,everyth:[31,39,46],evil:43,ewm:34,exactli:45,exampl:[15,25,37,38,40,43,44,46],except:[8,11,37,38,43,44],excess:38,execut:[11,12],exist:[2,6,9,11,17,19,36,43,44,46],exit:[12,40],expand:[2,46],expand_arrai:2,expand_on_rec:2,expect:[11,37,38,43,46],experi:[40,42,45,46],experiment:[40,42,46],explain:43,explan:36,explicit:[6,37],explicitli:[2,10,11,12,37,38,43],exposur:11,exposuretim:11,extend:[40,42,43,46],extens:[8,11,43,46],extern:[2,11,35,45,46],extra:[38,46],extract:[37,38,43,44],extran:25,extraordinarili:11,extrem:38,face:[2,41,42],fail:37,fakekm:40,fall:[12,38,43,46],falling_edg:12,fals:[2,6,8,10,11,12,14,15,17,21,24,25,28,29,30,31,34,38,43,44],far:46,faster:[2,14,40],fastest:40,favor:21,fbdev:40,fdisk:40,featur:[11,35,46],feed:3,feel:46,few:[10,37,38,40,41,43,44],ffmpegwrit:11,field:[2,14,30,44,46],file:[2,3,6,10,11,15,17,25,34,37,39,40,43,44,45,46],file_block:[2,3],file_db:8,file_nam:8,filehandl:[2,3],filenam:[6,11],filenod:6,filestructur:6,filesystem:40,fill:[23,44],filter:[25,38,46],final_recipi:37,find:2,find_packag:46,finish:[3,11,30,43],finish_off:12,finit:[11,42,46],firmwar:11,first:[21,23,28,30,37,43,44,46],first_param:27,fit:[12,43],fix:[6,37,40,43,46],fixat:43,flag:[2,3,6,23,37,40,43],flash:[12,30,31,37,38,43],flash_block:38,flash_l:[30,31,43],flatten:[6,46],flexibl:[10,37,40],flicker:46,flir:[11,37],float32col:[6,43],float64col:43,flush:6,flush_curr:6,fmform:[17,19],follow:[3,6,11,17,19,21,40,43,46],forc:[30,44,46],force_to:2,forev:40,form:[8,11,12,17,46],formal:[27,38,39,46],format:[2,3,6,21,37,43,44,46],formatt:[2,3,46],formdefault:17,forth:46,forthcom:42,forum:37,forward:2,found:[11,37,42,44],fps:[11,13,19,41],fraction:8,frame:[2,11,13,17,23,37,40,45,46],frame_trigg:11,framer:[11,13],framework:[43,45,46],free:[26,44],free_wat:26,freeli:[44,46],freetyp:40,freez:46,freq:43,frequenc:[12,21,25,37,38,43,44],frequent:37,from:[2,3,6,8,10,11,12,13,15,17,19,21,23,28,29,37,38,40,41,42,43,44,45,46],fromstr:11,full:[6,9,11,12,34,35,37,46],fuller:46,further:[2,11,37,43,46],futur:[2,3,6,8,21,37,38,39,40,43,44,45,46],gabor:43,gain:13,gap:[25,37],gap_zero:25,garbag:38,gather:43,gauss:13,gave:43,gdc:12,gener:[2,6,10,11,12,15,19,21,23,24,25,31,37,38,41,42,43,44,46],genicam:11,genotyp:44,get:[2,3,6,10,11,13,15,29,31,37,38,40,43,46],get_current_tim:37,get_ip:[2,3],get_nam:10,get_step_histori:6,get_stream:[2,11,37,45],get_sync_tim:37,get_thi:37,get_timestamp:[2,6],get_trial_data:6,get_weight:6,getattr:43,getter:[37,46],getvalu:11,gevent:40,ghasp:46,git:[6,15,19,37,40,41,46],git_vers:15,github:[12,35,37,40,46],gitlab:40,gitpython:40,give:[2,3,21,30,37,38,40,43,44,46],given:[2,3,6,8,10,11,12,14,21,25,29,37,38,40,43],glass:46,glfw:40,glitch:40,global:[15,46],gnueabihf:40,goal:[37,39,46],god:46,going:43,gonna:43,gonogo:[26,43],good:[37,38,40,43,46],googl:37,got:[37,40],govern:39,governor:40,gpio:[3,9,10,17,27,28,30,31,37,40,43,45,46],gpu:40,grab:[11,40],grad:34,grad_data:34,grad_list:29,graduat:[6,26,34,42,44],gram:6,grammar:46,graphic:[40,43,46],grate:43,green:[12,43],griddata:13,ground:13,group:[3,6,8,21,37,40],group__timefunct:46,group_nam:21,guarante:[12,46],guess:10,gui:[0,2,5,19,37,41,42,43,46],guid:[38,45],gyro:13,gyro_scal:13,gyroscale_2000dp:13,gyroscale_245dp:13,gyroscale_500dp:13,gyroscop:[13,37],h5f:6,had:[37,43,46],half:38,hand:38,handl:[2,3,10,11,12,23,37,38,43,46],handle_listen:2,handle_trigg:[10,14,31,43,46],handler:[2,3,46],handshak:[2,3,44],happen:[8,37,46],hard:[37,46],hardcod:[15,37,43,44],hardwar:[2,3,9,11,12,13,14,17,27,28,30,31,37,40,42,44,45,46],hardwareform:17,harshli:46,has:[2,6,8,10,11,14,21,25,28,37,38,40,43,44,45,46],hash:[6,15,46],hash_tabl:6,hasn:11,have:[2,6,10,11,12,21,27,28,29,30,37,38,39,40,41,42,43,44,45,46],haven:[2,37],hciuart:40,hdf5:[6,34,41,44,46],hdf:6,header:40,headless:40,heartbeat:46,heavili:42,height:[11,43],held:[11,38],help:[17,38,40],here:[2,6,11,12,15,17,19,29,38,40,43,44],hidden:46,hierarch:46,hifi:40,hifiberri:[40,46],high:[3,12,31,37,38,40,44],higher:40,highli:[11,37,43],histori:[6,21,43,44],history_t:6,hit:2,hold:[2,11,15],hole:44,home:11,honor:46,hope:37,horizont:[11,43],hover:46,how:[2,3,11,12,21,31,38,40,43,44,46],howev:[6,12,44],howto:46,html:[5,12,46],http:[5,8,11,12,13,35,36,37,40,42],human:[11,43,46],humbl:25,i2c1:40,i2c:[9,10,12,37,40],i2c_9dof:[13,37,45],i2c_arm:40,i2c_arm_baudr:13,i2s:40,i32:28,ideal:[12,40,46],ident:[2,43,44],identifi:[2,3,11,40,43],idl:[3,11],ight:[43,44],ignor:[3,6,11,29,40],iidc:11,iis:11,illustr:38,imag:[11,40,46],image1:11,image_path:11,imageitem:37,imageitem_timedupd:37,img:40,img_opt:11,immatur:[43,46],immedi:43,implement:[2,3,6,8,10,14,18,21,25,29,38,43,46],implicit:46,implicitli:[37,44],imposs:46,improv:[37,39],incap:46,includ:[2,3,6,9,11,12,37,43,44,45,46],include_baselin:6,inclus:6,incom:2,incompat:46,incomplet:[45,46],incorpor:[21,39],incorrect:[21,29,30,37,46],incorrectli:[21,37],increas:[6,13,37,40],increment:37,indefinit:[11,28],independ:[2,10,46],index:[2,11,25,45,46],indic:[2,11,40,46],indicipher:46,individu:[11,12,21,37,43,46],ineffici:46,infer:37,info:[6,12,37,43],inform:[2,6,11,12,37,44],infrastructur:[37,43,46],ing:2,inherit:[10,12,25,37,43,45,46],init:[2,11,12,15,40,43],init_audio:3,init_cam:[11,13],init_hardwar:[31,37,43],init_log:[2,3,10,37],init_manag:[21,43],init_network:[2,10,11,37],init_pigpio:[3,12],init_sound:[21,25],init_sounds_group:21,init_sounds_individu:21,initi:[2,3,10,11,12,15,21,23,28,37,38,40,44,46],initialis:[17,19],inner_kei:[2,37],input:[2,10,11,12,14,30,40,41,43],inspect:[38,46],instal:[11,13,37,41,44,45,46],install_opencv:11,install_pyspin:11,instanc:[2,11,38,40,43],instanti:[2,3,10,11,12,21,25,31,38,43,46],instead:[2,3,11,12,17,43,46],instruct:[39,46],int32col:[6,28,30,31,43],int_to_float:25,integ:[6,11,25,37,43],integr:[37,45],integrate_dur:[14,43],integrate_fram:13,intend:[2,37,45],intens:45,inter:3,interact:[11,40,46],interest:[37,38],interfac:[23,37,38,40,46],interfer:13,intermediate_node_1:37,intermediate_node_2:37,intern:[2,10,12,21,40,43],internet:44,interpol:[13,37],interpolate_fram:13,interrupt:37,interv:[3,43],intim:46,intra:46,introduc:[42,44],intuit:[38,46],invalid:11,invers:[10,12],inverse_pull_map:12,inverse_trigger_map:12,invert:[12,38],invoc:8,invok:[8,39],ioloop:[2,43],ipv4:[2,3],irrelev:46,is_al:8,is_set:43,is_trigg:[10,12],isdescript:[6,28,30,31,43],isdescriptor:6,isn:11,isoformat:[2,6,11,12,13,28,30,37,43],issu:[24,35,40,45,46],item:[38,43],items:[6,28,30],iter:[21,28,31,38,43],itertool:[2,12,23,28,29,30,43,46],iti:3,its:[2,3,6,10,11,12,17,19,21,25,37,38,39,40,43,44,46],itself:[6,11,12,23],jack2:40,jack:[3,17,23,25,37,40],jack_client:23,jack_last_frame_tim:46,jack_sound:[25,37],jackaudio:[40,46],jackclient:[3,20,22,37,40,46],jackd:[3,17,23,37,40],jackdstr:[3,17],jitter:46,joblib:46,joi:46,join:[2,8,19,41],jonni:[15,25],journalctl:44,json:[2,3,6,15,19,38,39,40,41,43,44,46],json_trick:40,juic:40,just:[2,11,13,21,23,28,36,37,38,40,43,44],keep:[2,3,6,13,21,37,40,43],kei:[2,3,6,8,10,12,25,40,43,44,46],kept:[2,12,23],keyboard:40,keyword:[8,17,19],kick:[40,46],kill:[2,38],know:[2,6,37,43,44],knowledg:11,known:[2,10,38,46],kwarg:[2,3,8,10,11,12,13,21,25,27,28,29,30,31,38,43,46],l_bandwidth:3,l_cal_port:3,l_cal_result:3,l_chang:2,l_child:[2,43],l_clear:[14,43],l_coher:2,l_confirm:2,l_continu:[2,37],l_data:2,l_file:2,l_forward:2,l_handshak:2,l_init:2,l_kill:2,l_measur:[14,43],l_noop:2,l_param:3,l_ping:2,l_process:27,l_start:[2,3,11],l_state:2,l_stop:[2,3,11,14,43],l_stopal:2,l_stream:[2,37],lab:35,label:46,lack:46,lambda:43,larg:[2,12,37],laser:43,last:[6,11,12,21,27,30],last_init_lock:11,last_opencv_init:11,last_stim:21,last_was_correct:21,latenc:[40,46],later:[6,11],latest:46,launch:[12,38,39,40,41,44],launch_pilot:[12,40],launch_termin:41,layer:[38,43,46],ldconfig:40,lead:11,learn:44,learnopencv:40,least:[2,11,27,37,46],leav:[38,40,43,44],led:[3,10,12,14,17,28,30,31,37,38,40,43,44,46],led_rgb:[3,12,17,28,30,40,43],left:[8,21,40,44],leftmost:44,len:[12,38],length:[38,43],less:[21,40,43,46],let:[37,38,40,43,44,46],letter:31,level:[12,19,23,25,31,34,37,41,46],levelnam:[2,3],lib:40,libasound:40,libdir:40,libffi:40,libhdf5:40,liblo:40,libportmidi:40,librari:[9,13,35,40,41,45,46],libreadlin:40,libsamplerate0:40,libsndfile1:40,libx264:11,libzmq:[40,41],lick:43,life:40,light:[28,30,31,38,43],lightli:9,like:[2,6,11,12,21,27,30,31,37,38,39,40,42,43,45,46],limit:[11,38,40,42,44,46],line:[11,12,17,19,40,43],lineag:[40,43],linear:21,linearli:21,link:[15,26,40,46],linux:[39,40],list:[2,3,6,8,11,12,21,27,29,30,31,33,34,37,38,40,43,44,46],list_opt:11,list_spinnaker_camera:11,list_subject:8,listen:[2,3,10,19,37,41,43],listen_port:2,lite:40,liter:46,literatur:44,littl:[23,43],live:2,load:[6,8,23,25,34,37,40,46],load_pilotdb:8,load_subject_data:34,load_subject_dir:34,local:[2,3,11,19,37,40,41],localhost:[2,11],locat:[3,6,19,41],lock:[6,11,23,43],log:[2,3,10,11,15,19,31,35,37,40,41,44,46],log_formatt:[2,3],log_handl:[2,3],log_level:[19,41],logdir:[2,3,10,15,19,41],logfil:46,logger:[2,3,10,31,46],logic:[10,11,12,21,37,38,40,42,43,45,46],login:40,logist:33,loglevel:[15,19,41],look:[3,40],lookup:3,loop:[2,11,23,37],loop_thread:2,lossless:11,lost:46,lot:[37,46],love:[37,46],low:[12,31,37,38,40,44],lower:[40,46],lsm9ds1:37,luigi:46,machin:[11,40,42,46],maco:[39,40,41],made:[2,6,8,11,37,38,46],mag_gain:13,maggain_12gauss:13,maggain_16gauss:13,maggain_4gauss:13,maggain_8gauss:13,magic:37,magnet:13,magnetomet:[13,37],magnific:45,mai:[8,13,37,42,44,46],main:[2,3,31,43,46],maintain:[12,38,43],mainten:41,major:[35,39,46],make:[2,3,6,8,10,11,14,17,19,21,37,38,39,40,42,43,45,46],make_dir:[17,19],make_punish:21,manag:[6,8,10,12,20,23,30,38,40,42,43,45,46],mani:[2,3,8,11,12,37,40,45,46],manpag:[3,17,40],manual:[6,39,41,46],manual_trigg:12,map:[3,8,10,11,12,17,21,28,29,31,40,43,46],march:[35,46],margin:37,mark:[30,31],mask:[3,17],mass:6,match:[2,43],materi:46,matur:46,max:[2,11],maximum:[12,40],mayb:44,mean:[8,11,13,21,37,43,46],meant:37,meantim:46,measur:[3,14,37,43],measure_evt:43,memlock:40,memori:[6,40,46],menu:[37,40,44],mere:44,merg:6,mesa:40,mesh:46,meshnet:46,mess:40,messag:[2,3,10,11,37,40,41,43,44,46],message_cont:37,messi:46,meta:45,metacalss:40,metaclass:[10,11,12,31,37,38,46],metadata:[42,46],meter:37,meth:12,method:[2,3,6,8,10,11,12,13,14,17,19,21,23,25,27,28,29,30,31,37,38,46],mice:[37,38,43],microsd:40,microsecond:[11,12,37,43,45],mid:3,might:[2,25,37,46],mild:30,millisecond:[11,12,43],mimc:46,min_siz:[2,11],mind:40,minim:[2,10,37],minimum:[6,12],minor:[35,46],minu:6,mirror:15,misalign:43,mlx90640:[9,13,37,45],mmap:40,mode:[6,11,12,14,21,30,38,40,43,46],model:[6,14,46],modif:[11,37,45],modifi:[6,9,13,21,37,40,46],modul:[3,12,15,17,19,21,23,24,34,37,38,40,45,46],moment:46,monitor:43,monocl:46,more:[10,11,12,21,37,38,40,42,43,45,46],most:[6,37,38,42,43],mostli:37,motion:[37,45],motiv:45,mous:[14,37,38,43,44],mouse_idx:[14,43],move:[14,29,42,43,44,45,46],move_dtyp:14,movement:[14,38,43],moviepi:40,mp4:11,msg:2,msg_counter:2,msgpack:40,msgport:[3,11,15,17,19,41,43],much:[37,38,43,45,46],muddi:10,multi:11,multihop:46,multipl:[2,3,11,12,23,37,38,42,43,46],multipli:[13,30],multiprocess:[2,11,23,46],must:[2,6,8,10,11,12,13,14,23,25,38,41,43],mutual:12,mysteri:37,n_click:3,n_rep:38,n_trial:[29,44],nafc:[26,44],nafc_gap:[26,30,37],name:[2,3,6,8,10,11,12,15,17,21,23,26,27,31,33,40,43,44,46],name_of_camera_in_task:27,nano:40,nativ:46,natur:44,nchannel:[3,17,23],nctl:40,ncurs:40,ndarrai:[2,11,13,23,25],ndtoverlai:40,ndtparam:40,neater:46,necessari:[10,31,37,40,43,46],necessarili:2,need:[2,3,6,10,11,12,14,17,21,25,29,31,37,38,39,40,43,44,46],needlessli:46,neither:40,nest:[17,19],net:43,net_nod:[2,3,10,11,37,43,45],network:[0,3,10,11,17,37,40,42,43,45,46],neurodata:46,neurosci:[37,45],neutral:[10,14],never:37,neverend:11,new_subject_fil:6,new_valu:6,next:[3,11,21,31,40,43],next_bia:21,next_stim:[21,43],nice:46,nmap:11,node:[2,3,6,10,11,37,40,43,46],nodemap:11,nois:[23,25,30,37,43],noise_amplitud:30,noisier:37,non:[37,38,43],none:[2,3,6,8,10,11,12,13,14,15,17,19,21,23,27,28,30,31,34,38,43],nonzero:30,noop:[27,43],normal:[8,12,37,40,43,44],nosepok:44,note:[2,11,12,39,40,43],noth:[8,25,43],notic:44,now:[13,21,23,30,31,37,38,41,42,43,45,46],nowtim:43,nperiod:40,npsappmanag:[17,19],npyscreen:[17,19,40,41,46],ntp:46,ntrial:29,nudg:37,num_stag:28,number:[2,3,6,8,10,11,12,13,17,21,23,28,29,30,31,37,38,40,42,43,45,46],numer:[8,11,43],numpi:[2,11,13,15,23,25,38,40,41,43],numpy_arrai:2,nut:10,object:[2,3,6,8,10,11,12,13,14,17,21,25,27,28,29,31,37,38,40,42,43,44,46],obnoxi:11,observ:46,obviou:[37,46],obvious:40,occur:[8,31],octob:35,odict:[31,43],off:[10,12,14,28,31,38,40,43,46],offroad:2,offset:37,offtim:38,often:[2,46],old:[6,37],oldest:46,omit:38,on_recv:2,onboard:40,onc:[2,11,12,31,37,40,43,44,46],one:[2,6,12,13,21,23,28,37,38,40,42,43,44,45,46],ongo:[42,43],onli:[2,6,10,11,12,23,27,31,37,39,40,43,44,46],onstart:[17,19],opac:24,opaqu:46,open:[2,3,6,11,12,28,30,31,37,40,43],open_dur:3,open_fil:3,open_hdf:6,opencl:46,opencv:[11,37,40,45,46],opencv_last_init_tim:11,openephi:46,opengl:40,oper:[8,10,11,27,37,39,40,41,42,43,44,45,46],opportun:40,oppos:38,optim:46,option:[2,3,8,11,37,38,39,41,43,44,46],order:[2,6,10,11,38,43,44,46],ordereddict:31,org:[36,42,46],organ:21,orient:37,origin:2,other:[2,6,10,11,12,13,15,21,25,27,37,38,40,43,46],otherwis:[2,3,6,11,12,21,23,31,43,46],our:[2,3,6,13,17,19,21,38,40,41,43,44,45,46],ourselv:3,out:[6,8,12,21,37,40,44,45,46],outbox:2,outcom:21,outport:23,output:[3,6,10,11,12,13,17,23,27,40,43,45],output_filenam:11,outright:46,outsid:[3,30,43],over:[2,11,12,13,17,19,21,28,29,31,37,40,42,43,46],overhaul:46,overlai:40,overlap:[11,37,38],overrid:[2,8,12,17,19,37],overridden:11,overrun:40,overview:45,overwrit:[11,38],overwritten:[11,12],own:[2,3,10,11,12,17,21,37,39,40,42,43,44,45,46],p16:[3,17,40],p75:[3,17,40],pack:2,pack_arrai:[2,11],packag:[23,38,40,41,46],pad:25,page:[37,42,45],pair:[2,37,43],panda:[6,8,33,40,41],panel:44,parallax:26,param1:27,param:[2,6,8,12,15,17,19,25,27,28,29,30,31,33,41,44,46],param_1:44,paramet:[2,3,6,8,10,11,12,13,14,15,17,19,21,23,24,25,27,28,29,30,31,33,34,37,38,40,42,43,44,46],parameter:[12,38,40,46],parent:[25,40,43,46],parent_pilot:43,parentapp:17,parentid:43,parentip:43,parentport:43,parlanc:12,parsimoni:46,part:[39,45],partial:6,particular:[6,10,11,21,44,46],particularli:[37,42,46],pass:[2,3,6,8,10,11,12,13,23,25,30,43,46],passag:43,passwd:40,password:[40,44],past:[6,29],past_protocol:6,patch:46,path:[2,3,6,11,15,17,19,25,34,40,41],pattern:46,pcm512x:40,pcm:40,per:[2,23,43,46],percent:[21,29],percentag:43,perfect:37,perform:[3,11,17,19,37,38,39,42,43,45,46],period:[2,40],permiss:[17,19],persist:42,phase:6,physic:10,pi3:40,pick:37,pickl:46,piec:[2,38,46],piecem:40,pig:[12,38],pigpio:[3,9,10,12,13,17,31,37,40,43,45,46],pigpiod:[10,12,38,46],pigpiomask:[3,17],pigs_funct:12,pillow:40,pilot:[0,2,10,17,19,28,37,38,39,41,42,43,46],pilot_button:46,pilot_db:[8,15,19,41,44,46],pilot_st:[2,3,43],pilotsetupform:[17,46],pimoroni:37,pin2:31,pin:[3,10,12,13,17,28,30,31,37,38,40,43],pin_bcm:12,pin_id:31,ping:2,pinout:10,pip3:41,pip:[40,46],pipe:23,pixel:37,place:[2,12,43],plai:[23,25,30,37,40,43],plaintext:37,plan:[37,45,46],play_continu:37,play_evt:23,play_punish:[21,43],playback:[40,43],pleas:[11,45],plot:[0,2,28,30,31,33,37,42,44,46],plot_list:43,plot_psychometr:33,plot_tim:2,plots_menu:37,plug:31,plugin:46,png:11,pngoption:11,point:[8,28,30,31,37,43,46],pointer:11,poissonian:3,poke:[3,17,28,30,31,38,40,43,44],polar:[12,37,38,40],poll:46,pool:[40,46],pop:2,pop_dialog:37,popul:[11,19,29,41,46],port:[2,3,11,12,17,19,21,23,28,30,31,40,41,43,44],port_calibr:12,port_id:31,port_nam:3,portal:46,pos:[6,28,30],posit:[10,37],possibl:[11,12,37,43,46],post:[6,46],postgresql:46,potenti:[37,46],practic:[37,43],pre:6,preambl:38,prebuilt:11,preciou:38,precis:[12,25,37,43,45,46],precomput:43,predict:46,pref:[2,3,6,10,11,12,19,23,25,31,37,38,39,40,41,44,45,46],prefdict:15,prefer:[12,15,37,40],prefix:43,prefs_fil:[3,15],prei:44,preinstal:39,prepar:[6,11,39,43],prepare_messag:2,prepare_run:6,prerequisit:39,present:[3,8,21,37,40,42,43,44,46],presetup:[39,41],presetup_pilot:[40,46],presetup_termin:41,press:[44,46],pretti:43,prettyprint:46,prevent:[37,38],previou:[6,12,37,38,43],previous:[37,38,40,46],primarili:[10,46],primit:43,principl:38,print:[11,44],prior:21,prioriti:[3,40],privileg:40,problem:[37,46],process:[2,3,11,23,37,38,39,43],produc:[37,40,43,46],product:[11,13],product_id:14,program:[44,46],programm:[17,19],programmat:43,progress:[3,11,45],project:46,prompt:[17,19,40],proper:37,properli:6,properti:[11,12,13,37,46],proport:[11,12,21,25,30,43],protocol:[6,12,15,19,26,29,37,41,42,46],protocol_nam:6,protocol_script:44,protocoldir:[6,15,19,41],provid:[2,11,12,37,41,42,43],psychometr:[32,37],psychopi:40,psychopy3:40,pub:46,publish:46,pull:[3,6,11,12,17,23,38,40],pull_map:12,pull_ud:38,pulldown:[3,12,17],pullup:[3,12,17,37],puls:[11,12,38,43],pulse_width:12,punish:[30,31,43],punish_block:[31,43],punish_dir:[30,31],punish_dur:[30,43],punish_stim:[30,31,43,44],purpos:[10,43,46],push:[2,40,46],push_id:2,push_ip:2,push_outbox:2,push_port:2,pusher:2,pushport:[3,15,17],put:[6,11,13,29,43],pwm:[12,40],pyglet:40,pyo:[3,17,24,25,40],pyo_serv:[3,24],pyo_sound:25,pyopengl:40,pyoserv:[3,20,22],pyparallel:40,pyper:46,pyqtgraph:[37,41],pyseri:40,pysid:[37,41,46],pyside2:[37,41,42],pyspin:[11,37],pytabl:[6,43],python2:40,python3:[40,41],python:[2,3,10,12,13,37,38,40,41,45,46],pyyaml:40,pyzmq:[40,41],q_lock:23,qt4:41,qt5:[37,41,42,46],qtcore:37,qtimer:37,queri:11,quest:37,queue:[2,6,11,23,37,43,46],queue_siz:11,quickli:37,quit:[23,40,46],quit_evt:[23,43],radio:46,raft:37,rais:[8,11,40],random:[21,43],randomli:[21,28,30],randomnumbersandlett:40,rang:[6,12,13,43],rapid:38,rare:21,rasbian:39,raspberri:[3,10,11,37,38,40,42,43,44,45],raspberrypi:40,raspbian:39,raspi:[12,37,40],rat:[12,37,40],rate:[2,3,17,21,23,27,37,40,43],rather:[2,9,11,12,23,25,37,38,40,43,46],rdisk:40,reach:30,read:[6,11,13,37,38,43,46],readabl:[11,43,46],readable_attribut:11,readi:43,readthedoc:46,real:12,realli:[2,40],realtim:[40,46],reason:40,reassign:[2,6],reboot:40,rebuilt:[37,45],receipt:37,receiv:[2,6,37,40,43,44],recent:[6,43,46],recipi:[11,37,46],recommend:[6,11,40,43],reconfigur:[40,46],reconstitut:2,reconstruct:6,record:[11,12,25,37],record_ev:12,recov:46,recurs:[17,19,38,46],recv:2,red:[12,31,43],redefin:[10,14,21,38],reduc:[37,40],redund:43,refactor:[6,37,38,45],refer:[5,11,12,29,30,31,43],reflect:[21,37,46],refrain:43,refresh:44,regimen:44,region:46,regist:[23,38,43,44],register_task:46,registri:46,regress:33,reinforc:30,rel:[2,6,11,25,27,37,44,46],rel_i:38,rel_x:38,relat:44,releas:[2,3,10,11,12,13,14,21,27,28,31,36,37,38,43,46],relev:[2,11,38,44,46],reli:[12,37,43,46],reliabl:[40,46],remain:[11,12,23,27,38,44,46],remov:37,renam:37,reorgan:37,repeat:[2,12,28,37],repeat_interv:2,repeatedli:[3,43],repetit:12,replac:[37,40,46],replic:[45,46],repo:[19,37,41],repo_dir:15,repo_directori:40,repodir:[15,19,41],report:[27,28,37,43,44,46],repositori:[6,9,15],repres:[8,12,46],represent:6,req_reward:[30,31,43,44],request:[2,3,6,11,30,31,37,44],request_reward:44,requir:[2,3,10,11,17,37,38,39,40,41,43,44,46],requirements_doc:37,requirements_termin:37,resampl:25,rescal:25,resembl:46,resend:[2,37],resent:2,reserv:8,reshap:40,resistor:[12,37,38],resiz:[40,46],resolut:[13,37],resourc:[10,11,14,37,38,43,46],respect:[8,11,12,37],respond:[2,21,30,43,44],respons:[21,28,30,31,33,40,43],rest:[2,3],restart:46,restructur:37,result:[3,12,27,37,43,46],resum:37,retriev:[11,12,44,46],return_id:27,returnthread:8,revers:[8,31,40,43],revis:11,reward:[12,28,30,31,37,40,43,44,45],reward_manag:44,rfs:[3,17,40],rgb:[12,38,40],rhythm:46,ride:[17,19],right:[8,21,23,40,44],rightmost:44,ring:13,rise:12,rising_edg:12,risk:[6,46],roadmap:45,robust:46,role:39,roll:[31,43,46],roll_span:34,roll_typ:34,roll_window:[30,31,43],rollmean:[30,31,43],root:[6,17,19],roughli:46,round:[38,43],rout:[2,46],route_port:2,router:[2,3,17,19,41,46],routin:[3,11,15,37,38,39,40,41,42,43,46],row:[3,29,43],rpi3:46,rpi:[12,40],rq_timestamp:[30,31,43],rsa:[40,44],rtprio:40,rtype:[2,11],rule:21,run:[2,3,6,8,10,11,12,17,19,23,31,37,38,40,41,42,43,46],run_task:3,runtim:[39,46],runtimeerror:8,rw_mode:11,ryanjdillon:15,s26:28,sadon:2,safe:[10,14],same:[2,6,8,11,12,13,21,37,38,40,43],sampl:[3,17,23,37,40,46],sarg:46,sargehttp:46,save:[2,3,6,11,37,38,43,44,46],save_data:6,scale:[13,14,18,30,45],scaling_governor:40,schema:[21,44],scheme:[10,46],scikit:41,scipi:[13,25,41],scl:13,scrappi:21,scratch:[43,46],screen:[40,46],screenshot:43,screwi:23,script:[10,11,12,13,17,38,39,45,46],script_:12,script_count:12,script_id:12,scroll:46,sda:13,sdk:[11,37,45],seamlessli:37,search:45,second:[8,11,13,37,43,44],second_group:21,section:[38,43,44,46],secur:40,sed:40,see:[3,5,6,10,11,12,13,17,19,30,35,37,38,39,40,41,42,43,44,45,46],seem:[11,46],segment:[30,31,43],select:[2,3,6,11,21,33,44,46],self:[6,12,21,30,31,38,43],semant:46,send:[2,3,11,12,37,43,44,46],send_outbox:2,sender:[2,27,37,46],sens:46,sensibl:45,sensor:[13,37,43,44,45],sent:[2,37,43],sent_plot:2,separ:[11,13,43,46],sequenc:38,sequenti:8,seri:[11,12,38,43],serial:[2,6,11,37,44],serializ:2,serialize_arrai:2,series_:12,series_thread:38,serv:[13,27,43],server:[3,17,23,24,25,40,46],server_typ:25,servic:[10,17,39,40,41,46],session:[6,37],set:[2,3,6,9,10,11,12,13,17,19,21,23,30,31,37,38,39,40,41,43,44,46],set_color:[38,43],set_l:[31,43],set_mod:38,set_pull_up_down:38,set_pwm_dutycycl:38,set_reward:[31,43],set_trigg:[21,43],set_weight:6,setattr:43,setter:[37,46],setup:[3,10,11,13,15,18,37,38,39,42,43,44,45,46],setup_mlx90640:13,setup_pilot:[3,10,12,16,40,46],setup_pilot_:44,setup_scal:16,setup_termin:[16,41,42],setupapp:[17,19],setupform:[17,19],setuptool:46,setvalu:11,setverbos:24,sever:[37,43],shade:43,shape:[6,11,13,28,30,37,42,44],shape_sensor:13,share:[42,46],shared_memori:46,shift:43,shitload:2,shock:46,shorter:46,shorthand:12,should:[2,3,6,8,10,11,12,15,17,19,21,23,25,27,29,30,37,38,39,40,43,44,46],shouldn:[12,43,46],shovel:37,show:12,shown:43,shut:[28,43],side:[21,44],signal:[2,3,6,11,14,23,28,30,31,37,38,43,46],silent:[23,25,37],similar:[6,37,38,43,46],similarli:[21,38],simpl:[6,11,37],simpli:43,simplifi:[2,37,38,41],simultan:6,sin:25,sinc:[11,31,37,38,40,43,46],sine:25,singl:[2,6,11,12,23,37,38,43,45],single_dur:38,sink:46,size:[37,43,46],skew:21,skvideo:11,sleep:[3,38],sleev:38,slight:37,slightli:40,slot:[37,46],slow:43,slower:40,small:8,smaller:46,smart:38,smarter:[2,46],smooth:39,smoothli:43,snd_rpi_hifiberry_dacplu:40,sndrpihifiberri:[3,17,40],sneaker:[12,37,40],sock:2,socket:[2,11,12,37],soft:40,softwar:[10,46],solenoid:[3,12,17,28,30,31,40,43,44],solidifi:43,solut:10,solv:46,some:[2,10,11,12,15,21,23,29,30,31,37,38,40,43,45,46],some_index_for_your_mous:38,someth:[40,46],sometim:37,somewhat:46,song:2,sorri:[38,40],soul:40,sound:[3,15,19,20,21,23,30,37,40,41,42,43,44,46],sound_dict:21,sound_info:43,sound_list:[21,25],sound_param:21,sound_stim:21,sound_widget:43,soundcard:40,sounddevic:40,sounddir:[2,15,19,25,41],soundfil:40,sourc:[2,3,6,8,10,11,12,13,14,15,17,19,21,23,24,25,27,28,29,30,31,33,34,46],sparkfun:13,spawn:[2,6,10,11,13,38,46],speak:40,speaker:[25,40,46],specif:[12,37,38,43,44,46],specifi:[2,3,8,12,28,31,37,38,40,43,44,46],speech:[25,44],speed:[37,40],spill:12,spinnak:[11,37,45],spiral:46,splash:3,split:[2,37,40,46],splitform:[17,19],sprawl:46,ssh:[40,44],stack:44,stackoverflow:8,stage:[3,6,12,28,29,30,31,42,46],stage_block:[3,27,28,30,31,43],stage_list:43,stage_nam:[27,28,30,31,43],stai:6,stamp:14,standalon:3,standard:[8,37,46],start:[2,3,6,8,10,11,14,15,17,23,24,27,28,29,30,36,37,38,40,43,44,46],start_now:27,start_plot_tim:2,startup:[40,46],stash:[2,6,11,43],stash_curr:6,state:[2,3,6,10,11,12,14,37,38,42,44,46],station:[2,3,40,43],stationari:23,statu:[2,6,12,21,37,40,46],stdout:[11,37],steadi:[14,43],step:[6,33,34,37,39,44,46],step_data:34,step_n:6,step_nam:[33,44],step_view:34,still:[8,11,38,40,42,43,46],stim:[21,25,30,42,43,44],stim_end:[30,43],stim_light:30,stim_manag:[21,37,43],stim_start:[30,43],stimuli:[21,30,37,40,43,44,45,46],stimulu:[21,30,31,37,42,43,44,45,46],stimulus_manag:44,stolen:[15,40],stop:[2,3,6,10,11,12,13,14,23,27,30,38,43,46],stop_evt:23,stop_run:6,stop_script:12,stopal:2,storag:[2,6,42,43,46],store:[3,6,12,13,15,19,21,23,31,38,40,41,43,44,46],store_group:21,store_script:12,store_seri:12,stored_color:38,str:[2,3,6,8,10,11,12,13,15,17,19,23,25,27,33,34,38],straightforward:38,stream:[2,10,11,27,37,40,45,46],streamlin:46,streamz:46,strict:38,strictli:[40,46],string:[3,6,8,11,12,17,25,26,29,37,38,40,43,44],string_of_camera_class:27,string_param:25,stringcol:[6,28,30,31,43],strongli:[40,44],structur:[6,14,21,35,38,40,41,43,44,45,46],stuck:[44,46],style:[0,34,37,46],stylesheet:[5,37],sub:[2,27,46],subclass:[2,8,11,21,25,40,43,46],subconfigur:43,subdevic:40,subdict:31,subdictionari:43,subject:[0,2,3,8,21,28,30,33,34,37,42,43,45,46],subject_id:33,subject_list:46,subject_nam:34,subject_posit:46,subject_protocol:33,submit:[23,45],submodul:[9,37],subset:34,substanti:37,success:12,sudo:40,suffici:10,suit:46,suitabl:11,sum:21,summar:34,summat:11,sunk:40,superclass:[11,12,13],superfast:37,support:[11,23,37,43,44,45],suppos:44,sure:[2,6,8,11,38,40],surprisingli:37,swarm:[40,42],synchron:[11,12,37,46],synthesi:40,sys:[11,40],system:[2,8,9,10,11,12,14,37,38,39,40,41,42,43,44,46],systemctl:40,systemd:[10,17,38,40,41],systemwid:[13,19,41],t2000:[3,17,40],tabl:[2,3,6,25,28,29,30,31,37,40,41,43,44,46],tableextens:[3,29],tag:[8,27,28,30,31,43,44],take:[2,10,11,12,38,40,43],taken:[8,10,38,43],target:[8,21,28,30,31,38,43],task:[2,3,6,8,10,12,14,17,27,28,29,30,37,38,40,41,45,46],task_class:3,task_list:[3,26,44,46],task_nam:6,task_param:3,task_typ:[3,43,44],taught:44,teach:44,tee:40,tell:[2,31,38,43],temperatur:[13,37,45],templat:42,tempor:44,temporari:37,temporarili:46,tensor2tensor:46,term:[44,46],termin:[0,2,3,8,11,17,19,23,28,33,37,39,40,42,43,44,46],terminal_ip:44,terminal_network:37,terminal_st:2,terminalip:[3,11,15,17],terminalport:11,terminalsetupform:19,terminolog:37,test:[3,11,26,37,39,40,43,45,46],test_dlc_lat:26,than:[2,9,10,11,12,21,23,25,37,38,40,43,46],thank:45,thei:[2,11,12,21,23,25,31,37,38,39,43,44,46],them:[2,3,6,10,12,21,37,38,40,42,43,44,46],thereof:8,thi:[2,3,6,8,10,11,12,13,14,17,19,21,23,25,29,31,34,37,38,40,42,43,44,45,46],thin:43,thing:[21,37,39,43,46],thingivers:46,think:37,thortex:46,those:[38,40,46],though:[2,43],thread:[2,3,6,8,10,11,12,13,28,30,31,37,38,43,46],threaded_color_seri:38,threaded_loop:2,threadgroup:8,threadsaf:10,three:[11,12,37,38,43,44],thresh:[14,21,27,43],thresh_trig:[14,43],thresh_typ:[14,43],thresh_val:14,threshold:[14,21,27,29,30,43,44],thresholded_linear:[21,30],through:[2,6,8,11,12,28,37,38,40,43,44,46],throughout:[37,46],throughput:37,thu:[2,11,37,39,40,42,43,46],thx:8,tick:[9,12,31,37,43,46],ticks_to_timestamp:37,time:[2,3,6,8,11,12,13,14,21,23,29,37,38,40,41,43,44,45,46],timeout:[8,12,40,43],timer:[2,38,43],timestamp:[2,3,6,9,10,11,12,13,14,28,37,38,43,45,46],to_csv:[6,37],toggl:12,token:25,told:[2,43],tone1:21,tone1_param:21,tone2:21,tone2_param:21,tone3:21,tone3_param:21,tone4:21,tone4_param:21,tone:[21,25,30,44],too:[10,40,46],tool:[34,40,45,46],tooltip:46,top:46,topolog:2,tornado:[2,40,41,43],torrent:40,tostr:11,total:[12,14],toward:37,tqdm:[11,41],trace:44,track:[43,46],train:[6,38,43,45],transfer:[3,40,46],transform:[8,12,26,27,46],transit:[12,37,38,42],translat:[37,46],transmiss:45,transmit:43,transpar:[38,43],treat:46,tree:[2,46],treelik:40,tri:[8,12],trial:[3,6,21,28,29,30,31,34,37,42,43,44,46],trial_count:[28,30,43],trial_data:[6,31],trial_end:[6,30,43],trial_manag:46,trial_num:[6,28,30,31,43],trial_view:32,trialdata:[6,28,30,31,43,46],trialwis:37,trick:38,trig_fn:21,trigger:[10,11,12,14,23,27,28,29,30,31,37,38,40,43,46],trigger_edg:12,trigger_fn:[10,14],trigger_map:[12,38,46],trigger_ud:38,trivial:[3,40,43],troubleshoot:45,ttl:[2,11,12],tup:33,tupl:[6,8,11,12,13,21,33,38],turn:[10,12,14,28,30,31,38,40,42,43,46],tweak:46,twice:13,two:[2,6,9,10,11,12,28,30,37,39,40,43,44],txt:[13,37,40],type:[2,3,6,8,10,11,12,13,14,15,17,21,23,25,27,28,29,30,31,33,37,38,40,43,44,46],typic:[2,3,6,10,12,14,19,33,37,38,41],ubuntu:39,uli:42,ultim:46,umount:40,unambigu:[37,46],unavail:46,unchang:37,unclear:46,uncompress:11,undecid:46,under:[40,42,43],underrun:40,undocu:[27,34],unfinish:34,unfold:[17,19],unfold_valu:[17,19],unfortun:43,unhandl:8,unicod:15,unif:44,unifi:[2,21,39,43,46],unimport:46,uniqu:[2,8,40,43],unit:[12,44,46],univers:[40,44],unix:39,unless:[2,6,31,46],unlik:[2,12],unload:[10,14],unmount:40,unmountdisk:40,unnecessari:25,unpack:[38,43],unpickl:46,unproduct:46,until:[2,8,11,21,23,27,37,40,43],unus:[11,23,25],unzip:40,upcom:27,updat:[2,6,12,14,19,21,29,30,35,38,40,41,43,44],update_biographi:6,update_dur:43,update_histori:6,update_protocol:37,update_st:3,update_weight:6,upgrad:[40,41,46],upon:[2,12,23],upstream:[2,3,17,40,43],upstream_ip:2,url:37,usb:[10,27,37,40,43],use:[2,3,6,10,11,12,13,17,19,33,37,38,40,41,42,43,45,46],use_histori:6,used:[2,3,6,10,11,12,14,17,19,23,25,26,28,30,37,38,40,41,42,43,45,46],user:[11,12,15,17,19,37,40,41,42,43,44,45],usernam:40,userpref:40,uses:[2,9,10,11,12,13,14,21,37,38,40,41,43,46],using:[2,3,12,13,14,21,25,30,37,40,43,44,46],usr:[3,15,17,19,40,41],usual:[6,10,33,43],util:[0,40],v4l2:11,v4l_info:11,val:11,valid:2,valu:[2,3,6,8,11,12,13,14,15,17,19,21,27,30,31,38,43,44,46],value_1:44,valv:[3,12],var_i:33,var_x:33,vari:12,variabl:[2,3,6,10,11,12,13,15,21,23,25,28,29,30,31,33,37,38,43,46],variat:30,vc4cl:46,vcc:13,vel:14,vel_tot:[14,43],veloc:[27,43],vendor_id:14,venv37:15,venv:15,verbos:8,veri:[10,37,38,40,45,46],versa:[12,25],version:[2,6,9,11,12,13,36,37,38,39,40,41,43,44,46],vertic:[11,37],via:[2,21,40,43,44],vice:[12,25],video4linux:11,video:[2,11,19,26,37,39,41,45,46],video_child:[26,27,37],video_writ:11,videocaptur:11,videox:11,vidioc_qbuf:11,view:[6,46],viola:43,virtu:44,virtual:46,vision:[11,37],visual:[19,37,40,41,42,43,45,46],visul:34,viz:[15,37],vizdir:[15,19,41],vol:[12,31],voltag:[12,43],volum:[3,12,31,44,45],vowel:25,waf:40,wai:[12,37,40,46],wait:[3,8,11,13,23,28,37,38,43],wall:44,want:[2,40,43,46],warn:[10,14,37],wasn:21,watchdog:12,water:[3,12,26,28,30,40,44],wav:[2,25],wave:25,wavfil:25,web:46,webcam:[11,37],wehr:35,weigh:14,weight:[6,21],weight_tabl:6,well:[12,37,41,43,44,46],were:[37,38,43],what:[2,3,6,10,11,12,21,30,31,36,37,38,43,46],whatev:40,wheel:[14,26,27,38,40],wheel_0:43,wheel_:43,wheel_child:[26,27,43],when:[2,3,6,8,10,11,12,13,14,21,23,27,28,30,31,37,38,40,43,44,46],whenev:[6,12,37,43,46],where:[2,6,8,21,27,37,43,44,46],whether:[6,8,11,12,14,21,28,30,34,38,40,43,46],which:[2,3,6,10,11,12,13,17,21,25,30,34,37,38,39,40,41,42,43,44,46],whistl:38,white:[25,30,37,43],whitepap:[37,40,42,43,45],who:46,whole:[6,37],wholli:45,whose:[8,12,43],widget:[5,17,19,44],widget_list:17,width:[11,12,38,46],window:[19,21,29,31,37,41,44,46],wintyp:40,wire:2,within:[2,39,40,43,44,46],without:[2,6,10,12,37,38,44,46],wizard:[40,46],won:[10,43],work:[3,10,11,13,14,43,45,46],workaround:11,world:3,worth:46,would:[2,6,8,12,37,46],wouldn:37,wrap:[12,37],wrapper:[11,12,40,43],writable_attribut:11,write:[6,11,12,40,42,45,46],writeabl:11,writeable_attribut:11,writer:[11,23],written:[11,37,38,43,46],wrote:37,wth:11,www:[11,13,36,40,42],x11:40,x264:[11,37],x_vel:43,xinit:40,xml:11,xorg:40,xserver:40,y_vel:43,yes:[38,40,44,46],yet:[11,18,37,43,46],yield:[21,43,46],you:[2,8,11,12,13,21,37,38,40,43,44,45],young:45,your:[11,40,44,45,46],your_disk:40,zero:[23,25,28,29],zero_arr:23,zeromq:[2,40,41],zip:38,zmq:[2,37,41,46],zyre:46},titles:["core","gui","networking","pilot","plots","styles","subject","terminal","utils","external","hardware","cameras","gpio","i2c","usb","prefs","setup","setup_pilot","setup_scale","setup_terminal","stim","managers","sound","jackclient","pyoserver","sounds","tasks","children","free_water","graduation","nafc","task","viz","psychometric","trial_viewer","Changelog","v0.2.0 (October 26, 2019)","v0.3.0 (March 3rd, 2020)","Writing a Hardware Class","Installation","Pilot Setup","Terminal Setup","Overview","Writing a Task","Training a Subject","Autopilot - Distributed Behavior","To-Do"],titleterms:{"3rd":37,"class":[38,43,46],"new":[37,45],The:43,Using:[43,44],addit:43,agent:43,attribut:43,audio:40,autopilot:46,beambreak:38,behavior:46,bonsai:46,bug:46,bugfix:37,camera:11,changelog:35,child:43,children:27,close:46,code:37,com:46,complet:46,connect:44,core:0,creat:44,creation:44,data:43,dblink:46,debug:44,deeplabcut:46,discrim:43,discuss:46,distribut:43,ephi:46,extern:[9,37],featur:37,forum:46,four:43,free_wat:28,googl:46,gpio:[12,38],graduat:29,group:46,gui:1,hardwar:[10,38,43],high:46,href:46,http:46,i2c:13,improv:[40,46],indic:45,initi:43,input:38,instal:[39,40],integr:46,jackclient:23,led_rgb:38,librari:37,loop:46,low:46,lowest:46,major:37,manag:21,manual:[40,44],march:37,medium:46,method:43,minor:37,modul:42,nafc:[30,43],network:2,octob:36,open:46,option:40,output:38,overview:42,param:43,parameter:43,perform:40,pigpio:38,pilot:[3,40,44],pipelin:46,plot:[4,43],pnfrhdtza:46,pref:[15,43],preinstal:40,prerequisit:41,prioriti:46,process:46,program:42,protocol:44,psychometr:33,pyoserv:24,rasbian:40,raspbian:40,reinforc:43,request:43,roadmap:46,run:44,script:[40,41],setup:[16,40,41],setup_pilot:17,setup_scal:18,setup_termin:19,smart:43,sound:[22,25],span:46,stage:43,step:40,stim:20,structur:[37,42],style:5,subject:[6,44],tabl:45,task:[26,31,42,43,44],termin:[7,41],todo:[2,3,6,11,21,25,38,39,40,42,43,44],topic:46,tour:42,train:44,trial_view:34,updat:37,usb:[14,38],user:46,util:8,veri:43,version:35,video:40,vision:46,viz:32,what:45,wheel:43,wizard:44,write:[38,43]}})
